Release oslo.log for 2023.1 Antelope
During 2023.1 Antelope cycle the upper constraint version bump patch
could not merge [1], as cross-nova-py310 failed. This means that during
the cycle everything was tested with oslo.log 5.0.0, so as we discussed
this during release meeting [2] we needed to revert the patches that
most likely cause the job failure. This release contains those reverts,
and clarifies what oslo.log version should be used in 2023.1 Antelope
release.

[1] https://review.opendev.org/c/openstack/requirements/+/873390
[2] https://meetings.opendev.org/meetings/releaseteam/2023/releaseteam.2023-03-17-14.00.log.html#l-133

Change-Id: I53885de1f413495d843a63019a419833fe8e9066