Fix tools/announce by using importlib.metadata instead of setuptools
With release of new setuptools running setup.py started
returining extra line. Eventually, running setup.py as a script is not
considered as good practice and we should not rely on it's output much.
Instead, we can leverage importlib.metadata as it should be more
reliable in long run.

Also it allow us to remove our previous workaround:
https://review.opendev.org/c/openstack/releases/+/879065

Change-Id: I52714ae92458efc0628091762cb77cf3ddb058f9