[tricircle] Transition Rocky to End of Life
As discussed in the mail thread [1][2], this patch transition the Rocky
branch to End of Life. The last patch of the branch will be tagged with
rocky-eol tag. stable/rocky branch cannot be used anymore and will be
deleted if this patch merges.

This is needed as stable/rocky is not actively maintained in recent
period, thus gates are mostly broken due to job failures. Besides,
by removing these branches, infra resources will be freed up, too.

Please try to identify any zuul job, that is defined outside of the
repositories in this patch (for example in openstack-zuul-jobs, etc.)
and won't be used anymore if stable/rocky is deleted. Propose a job
removal patch for them.

Please +1 if the team is ready for us to proceed with this transition,
or -1 if there are still some activity on the branch and the team wants
to continue to maintain it.

[1] https://lists.openstack.org/pipermail/openstack-discuss/2023-January/031922.html
[2] https://lists.openstack.org/pipermail/openstack-discuss/2023-April/033386.html

Change-Id: I3e7c2a9f568605d3b5cb7a2c6e1bf79b1d536fa3