Release cliff for Bobcat-1 milestone
This is a library release for cliff for the Bobcat-1
milestone. This repository contains commits that have not been releases
for this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 4.2.0..72e81d7
72e81d7 Removing helper functions providing Python < 3.3 compatibility

Change-Id: Icac0f591e155281fd8cc1bac8b72dc8fa01f8d00
Signed-off-by: Elod Illes <elod.illes@est.tech>