Release ironic-lib for Bobcat-1 milestone
This is a library release for ironic-lib for the Bobcat-1
milestone. This repository contains commits that have not been releases
for this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 5.4.0..853bedc
853bedc Upgrade to latest hacking - v6
7a00c0d Exclude all files starting with . from flake8 tests
ec6ba65 Use direct writes when erasing the GPT

Change-Id: Ieaa71c51a21390b38908ae3dc38e4afd76610ed0
Signed-off-by: Elod Illes <elod.illes@est.tech>