Release keystoneauth for Bobcat-1 milestone
This is a library release for keystoneauth for the Bobcat-1
milestone. This repository contains commits that have not been releases
for this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 5.1.2..2e40bbf
2e40bbf Update master for stable/2023.1
44e5b2d New auth plugin v3oidcdeviceauthz
ca28df8 OAuth 2.0 Mutual-TLS Support
192a27a Remove six
bef3dfb setup.cfg: Restore python_requires after a typo
737790f Allow federation to work with unversioned auth_url
19db34e Switch to 2023.1 Python3 unit tests and generic template name
aaca2b3 Update master for stable/zed
be2223c Replace abc.abstractproperty with property and abc.abstractmethod

Change-Id: Ie82802dd77a0cbf8a7de42939a434acc8f054713
Signed-off-by: Elod Illes <elod.illes@est.tech>