Release keystonemiddleware for Bobcat-1 milestone
This is a library release for keystonemiddleware for the Bobcat-1
milestone. This repository contains commits that have not been releases
for this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 10.2.0..92cdf8a
92cdf8a Add timeout for requests
a59020f OAuth 2.0 Mutual-TLS Support
1c820f0 Add missing doc requirements

Change-Id: I228ab7f026f2bb3d7fd950c6fca9c0aceae1b02a
Signed-off-by: Elod Illes <elod.illes@est.tech>