Release osc-lib for Bobcat-1 milestone
This is a library release for osc-lib for the Bobcat-1
milestone. This repository contains commits that have not been releases
for this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 2.7.0..db9cdc9
db9cdc9 setup.cfg: Replace dash-separated config
6152b68 Add 'SizeColumn' formatter
d0d30b3 Update master for stable/2023.1
12900da tox 4 compatibility
1deaaec formatters: Cast *all* columns before outputting

Change-Id: Icaf9d59113d87c73fec9190982d013a871fd708e
Signed-off-by: Elod Illes <elod.illes@est.tech>