Release oslo.messaging for Bobcat-1 milestone
This is a library release for oslo.messaging for the Bobcat-1
milestone. This repository contains commits that have not been releases
for this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 14.2.0..fd2381c7
fd2381c7 Disable greenthreads for RabbitDriver "listen" connections
9b1e2dc4 Update master for stable/2023.1
115cfb5b Fix typo in quorum-related variables for RabbitMQ

Change-Id: Ic67679a03512ccf4599d2f5aa615b0877e4795c6
Signed-off-by: Elod Illes <elod.illes@est.tech>