Release python-adjutantclient for Bobcat-1 milestone
This is a library release for python-adjutantclient for the Bobcat-1
milestone. This repository contains commits that have not been releases
for this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 0.12.0..2f405f1
db795e2 Fix listing and removing roles from users.
c50e3fe Fix typo in help text
5a92d90 Bugfix: Six removal introduced regression in v1/users
b4ef3e6 Add check-requirements to Zuul tests

Change-Id: Ie9e32f8bd44e4c3b3486a572bb2c48bef1ed42d6
Signed-off-by: Elod Illes <elod.illes@est.tech>