Release python-heatclient for Bobcat-1 milestone
This is a library release for python-heatclient for the Bobcat-1
milestone. This repository contains commits that have not been releases
for this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 3.2.0..668d85f
668d85f Add Python 3.10 to supported runtime.
e2c006b Update master for stable/2023.1
3edca52 Switch to 2023.1 Python3 unit tests and generic template name

Change-Id: I05a175b28bb0898891c62c125ec8a59bc195f7e8
Signed-off-by: Elod Illes <elod.illes@est.tech>