Release python-ironicclient for Bobcat-1 milestone
This is a library release for python-ironicclient for the Bobcat-1
milestone. This repository contains commits that have not been releases
for this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 5.1.0..c2f1573
c2f1573 Update dependencies for pep8: to hacking v6
e914086 Accept configdrive as a JSON file
138d797 Update master for stable/2023.1

Change-Id: Iaa68a6150645715c25c53f9ca2565e266f61ea13
Signed-off-by: Elod Illes <elod.illes@est.tech>