Release python-neutronclient for Bobcat-1 milestone
This is a library release for python-neutronclient for the Bobcat-1
milestone. This repository contains commits that have not been releases
for this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 9.0.0..d342171
52653c9 Remove the CLI code from the Neutron client.
0b271d3 OSC: Remove BGP calls to neutronclient
3071dfb Update master for stable/2023.1

Change-Id: If28383955fd8ccd1c8b2c8ca8f15edd4eed2faac
Signed-off-by: Elod Illes <elod.illes@est.tech>