Release stevedore for Bobcat-1 milestone
This is a library release for stevedore for the Bobcat-1
milestone. This repository contains commits that have not been releases
for this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 5.0.0..2d99ccc
cc8f04e Revert "Moves supported python runtimes from version 3.8 to 3.10"
da129b1 Moves supported python runtimes from version 3.8 to 3.10
43a79dd Update master for stable/2023.1
462be04 Catch NotADirectoryError error

Change-Id: I0a5de85554044658334bd6c93882646173f41589
Signed-off-by: Elod Illes <elod.illes@est.tech>