Release sushy for Bobcat-1 milestone
This is a library release for sushy for the Bobcat-1
milestone. This repository contains commits that have not been releases
for this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 4.4.2..f03d77e
eee778a Retry on ilo state error
a1eef45 Update pep8 dep hacking to latest: v6
c416bf1 Handle TransferMethod in vmedia insertion
2506b87 Add TransferProtocolType for any general error that mentions it
a59cf6c Exclude all files starting with . from flake8 tests
4a40143 Update master for stable/2023.1
fbf44e9 Handle non-default language for registries

Change-Id: Ie9b9ac25c22c2156a1ba345cf116d95220a05d8c
Signed-off-by: Elod Illes <elod.illes@est.tech>