[cinder] transition train to end-of-life
As discussed at the cinder weekly meeting on 2023-05-31 [0], the
cinder project wishes to declare the stable/train branch of all
branched deliverables EOL and requests that the release team delete
the branch.

The stable/train branch of cinderlib was previously made EOL and
deleted by change I25d78267ab4a.

[0] https://meetings.opendev.org/meetings/cinder/2023/cinder.2023-05-31-14.00.log.html#l-32

Change-Id: I0326e54bb50ceb03e5c65f225c4765467ad4b3e7