Release a patch version of keystoneauth
last releast of keystoneauth is actually broken, because it drops six in
requirements, but there is still one single plugin missed in the
cleanup. This leads to an error of missing six for every sdk/cli/ansible
users that are just installing openstacksdk (and pulling latest
keystoneauth1). The issue itself was fixed in
https://review.opendev.org/c/openstack/keystoneauth/+/883383 and we need
to release it now.

Change-Id: Ifa7288f2cad259c0569979d7744dd944c1165571