Bobcat-2 release for keystonemiddleware
This is the Bobcat-2 milestone release for keystonemiddleware.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If the team is happy with the current release hash, please +1 this
patch and we will process the release. If you need to get some
important changes merged first, leave a -1 and update the patch with
a new commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed after July 6th.

Change-Id: I0bee8a1e7285d971dc7c6ca5d09f25271848d993
Signed-off-by: Elod Illes <elod.illes@est.tech>