Release horizon 23.2.0(bobcat)
This patch cut the first release of the horizon for the bobcat cycle.
horizon recently bumped the upper version of Xstatic-jQuery version from
1.12.4.1->3.5.1.1[1] but the patch which update the upper bound in the
openstack/requirements is failing as of now [2] beacuse horizon 23.1.0
depends on XStatic-jQuery<2 and >=1.12.4.1. The main purpose of
this release is to fix the above version compatibility  issue.

[1] https://review.opendev.org/c/openstack/horizon/+/887548
[2] https://review.opendev.org/c/openstack/requirements/+/887933

Change-Id: I5ce388ff43e4ee951e56a2461cc1a959dde64426