OpenStack 2024.1 Caracal release schedule
Here is the proposed release schedule for OpenStack 2024.1 "Caracal".

This is a 26-week cycle with a final release date on April 3, 2024.
It has a longer milestone-2 period to accomodate Thanksgiving and
end-of-year holidays.

Also fix an issue in Bobcat release schedule where Antelope cycle-trailing
deliverables were mentioned as Bobcat cycle-trailing deliverables.

Change-Id: I0fb0e61e79e7e4c51d42cabd7c05bb15a05275fa