Release final automaton for 2023.2 Bobcat
This proposed a release for automaton to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 3.1.0..9255778
9255778 Update master for stable/2023.1

Change-Id: I9c90a1a42a18725151288ba13ee4192b48f911a0
Signed-off-by: Elod Illes <elod.illes@est.tech>