Release final ceilometermiddleware for 2023.2 Bobcat
This proposed a release for ceilometermiddleware to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 3.1.0..10e3d15
10e3d15 Add 'check_requirements' to CI

Change-Id: I752557ff0f6cac6845b87ac207db6a8cba892371
Signed-off-by: Elod Illes <elod.illes@est.tech>