Release final heat-translator for 2023.2 Bobcat
This proposed a release for heat-translator to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 2.7.0..bb13958
bb13958 (HEAD -> master, origin/master, origin/HEAD) Use upper-constraint in doc generation
2becf4b (gerrit/master) Fix passenv for tox4

Change-Id: Ib97ce0f389e6c9e28133f59bf557426303343662
Signed-off-by: Elod Illes <elod.illes@est.tech>