Release final keystonemiddleware for 2023.2 Bobcat
This proposed a release for keystonemiddleware to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 10.4.0..d36c86c
fe644ed Imported Translations from Zanata
cbf5969 Switch to 2023.1 Python3 unit tests and generic template name

Change-Id: Id63b909f2d8c793c5d86ddd99c185550c4724186
Signed-off-by: Elod Illes <elod.illes@est.tech>