Release final kuryr for 2023.2 Bobcat
This proposed a release for kuryr to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 2.7.0..bc7e8e3
bc7e8e3 Use py3 as the default runtime for tox
c11675d Update master for stable/2023.1

Change-Id: If2231b017543aa9101ba5a4797e4ca64b4e7c0ee
Signed-off-by: Elod Illes <elod.illes@est.tech>