Release final metalsmith for 2023.2 Bobcat
This proposed a release for metalsmith to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 2.0.0..968c9a0
968c9a0 Remove CentOS Stream 8 jobs
a3b966d Add a CentOS Stream 9 bios job
b63793d Use jammy nodes to run CI jobs
97b95f1 Add centos9 based job

Change-Id: I2f5264b77374a76047eab4657741ab2d7cd947f9
Signed-off-by: Elod Illes <elod.illes@est.tech>