Release final mistral-extra for 2023.2 Bobcat
This proposed a release for mistral-extra to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 13.0.0..2e0139b
9e2ba46 Fix Nova client version discovery
3385694 Add monitoring server to mistral-extra
44b3e5d Fix test suite bug with barbican client

Change-Id: Ie03e1351938f3b85781a7effd78a8bee83319e84
Signed-off-by: Elod Illes <elod.illes@est.tech>