Release final mistral-lib for 2023.2 Bobcat
This proposed a release for mistral-lib to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 2.8.0..7bbfa4f

Change-Id: I7d4a6dbf0964ac9d85ce62a4208a38ee68f65e31
Signed-off-by: Elod Illes <elod.illes@est.tech>