Release final monasca-common for 2023.2 Bobcat
This proposed a release for monasca-common to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 3.7.0..cdc03f9
cdc03f9 Fix zuul config error
1ca4152 Fix deprecated zuul queue syntax

Change-Id: Ifa678c016fb37c3bd5059af87fea4c20ae4ab1cf
Signed-off-by: Elod Illes <elod.illes@est.tech>