Release final os-ken for 2023.2 Bobcat
This proposed a release for os-ken to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 2.6.0..41c852e4
a87d983e Add Python 3.10 support to setup.cfg
8cd03eb2 Remove superfluous sudo
760bfb58 Cleanup py27 support
6f285a65 Add releasenote for quagga fix
17570b4c Fix quagga container
e9dc289d Make neutron-dynamic-routing job non-voting
7e280fdc Update master for stable/2023.1
6dfd5da3 Docs build: Add requirements.txt to deps
5f986ac1 Fix tox.ini for tox4

Change-Id: Iafd184cc230836d152342e3e48db8bc05f1486b9
Signed-off-by: Elod Illes <elod.illes@est.tech>