Release final osc-lib for 2023.2 Bobcat
This proposed a release for osc-lib to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 2.8.0..2f5167f
1a6c1c2 Revert "Do not output log on bash autocomplete generation"

Change-Id: I29351ded20afac0575279766a0f2d622c28b9496
Signed-off-by: Elod Illes <elod.illes@est.tech>