Release final oslo.cache for 2023.2 Bobcat
This proposed a release for oslo.cache to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 3.4.0..06f76e5
06f76e5 Imported Translations from Zanata
8d856e1 Bump bandit
7a596b7 Imported Translations from Zanata

Change-Id: I8321f8845a78c807b38827fe386c0e27483055e3
Signed-off-by: Elod Illes <elod.illes@est.tech>