Release final oslo.concurrency for 2023.2 Bobcat
This proposed a release for oslo.concurrency to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 5.1.1..1abc8e0
b4aec26 Bump bandit
9c7ecb4 Imported Translations from Zanata
28fd71f Revert "Moves supported python runtimes from version 3.8 to 3.10"
d41e8f3 Moves supported python runtimes from version 3.8 to 3.10
8c582a4 Update master for stable/2023.1

Change-Id: I2cdf40fd017fa238b6c7216fea23b9a461751a0e
Signed-off-by: Elod Illes <elod.illes@est.tech>