Release final oslo.config for 2023.2 Bobcat
This proposed a release for oslo.config to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 9.1.1..28187da
3dff974 Bump bandit and make oslo.config compatible with latest rules
44725fa Revert "Moves supported python runtimes from version 3.8 to 3.10"
32d5549 Moves supported python runtimes from version 3.8 to 3.10
1654dcc Update master for stable/2023.1

Change-Id: I0db781cd9ab3de2c08526f8a06f03ca074d1b1f2
Signed-off-by: Elod Illes <elod.illes@est.tech>