Release final oslo.context for 2023.2 Bobcat
This proposed a release for oslo.context to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 5.1.1..5168ccd
5168ccd Revert "Moves supported python runtimes from version 3.8 to 3.10"
2a3c812 Moves supported python runtimes from version 3.8 to 3.10
e17e485 Update master for stable/2023.1

Change-Id: I4cd6e6d8a75960d70ded9a371fa05ee6900064e7
Signed-off-by: Elod Illes <elod.illes@est.tech>