Release final oslo.i18n for 2023.2 Bobcat
This proposed a release for oslo.i18n to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 6.0.0..e356fe9
e356fe9 Imported Translations from Zanata
a74cbd9 Imported Translations from Zanata
47d6d34 Revert "Moves supported python runtimes from version 3.8 to 3.10"
14ac5fa Moves supported python runtimes from version 3.8 to 3.10
567dd05 Update master for stable/2023.1
a90c449 tox: Remove 'ignore_base_python_conflict'

Change-Id: I19dd497ca227df83334f702b6d228d5a048d482e
Signed-off-by: Elod Illes <elod.illes@est.tech>