Release final oslo.limit for 2023.2 Bobcat
This proposed a release for oslo.limit to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 2.1.0..c9ee0ad
c9ee0ad Bump bandit
fc64349 Revert "Moves supported python runtimes from version 3.8 to 3.10"
886e717 Moves supported python runtimes from version 3.8 to 3.10
de69de1 Update master for stable/2023.1

Change-Id: I5169efcc9244d825171bde286f8af25ec6bc9ac8
Signed-off-by: Elod Illes <elod.illes@est.tech>