Release final oslo.log for 2023.2 Bobcat
This proposed a release for oslo.log to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 5.1.0..82a59af
bf52b05 Bump bandit
eaf6089 Imported Translations from Zanata
a0c3ca8 Revert "Moves supported python runtimes from version 3.8 to 3.10"
6abf69e Moves supported python runtimes from version 3.8 to 3.10
ecfe66f Update master for stable/2023.1
15c71dd Imported Translations from Zanata
2b6f580 Cleanup py27 support

Change-Id: I568be1190433321d65e2ccdbfd4488d96966ea43
Signed-off-by: Elod Illes <elod.illes@est.tech>