Release final oslo.metrics for 2023.2 Bobcat
This proposed a release for oslo.metrics to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 0.6.0..5b47840
5b47840 Bump bandit and make oslo.metrics compatible with latest rules
40c1def Revert "Moves supported python runtimes from version 3.8 to 3.10"
e01117f Moves supported python runtimes from version 3.8 to 3.10
d399b28 Update master for stable/2023.1

Change-Id: I64551cbf840d4e9fdaf2675d95aad3e5b51ba5aa
Signed-off-by: Elod Illes <elod.illes@est.tech>