Release final oslo.middleware for 2023.2 Bobcat
This proposed a release for oslo.middleware to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 5.1.1..4ba32ed
207541f Use upper-constraint in doc generation
4153ffe Imported Translations from Zanata
ad9c204 fix spelling
2eb512c Revert "Moves supported python runtimes from version 3.8 to 3.10"
089fa36 Moves supported python runtimes from version 3.8 to 3.10
170e35a Bump bandit and make oslo.middleware compatible with latest rules
05be604 Update master for stable/2023.1

Change-Id: I31c4e7ed9bfe7c295d3eb551f554eb2c523313a2
Signed-off-by: Elod Illes <elod.illes@est.tech>