Release final oslo.policy for 2023.2 Bobcat
This proposed a release for oslo.policy to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 4.2.0..37de6f3
37de6f3 Imported Translations from Zanata

Change-Id: Ie2e555b07de58078495c6ee4ae1ea50b962ff289
Signed-off-by: Elod Illes <elod.illes@est.tech>