Release final oslo.reports for 2023.2 Bobcat
This proposed a release for oslo.reports to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 3.0.0..ec934c2
ec934c2 Imported Translations from Zanata
61ce374 Revert "Moves supported python runtimes from version 3.8 to 3.10"
a1b8dc2 Moves supported python runtimes from version 3.8 to 3.10
15bfc2c Update master for stable/2023.1

Change-Id: I9902dd6744011243daea2eaa0307320727ba3791
Signed-off-by: Elod Illes <elod.illes@est.tech>