Release final oslo.rootwrap for 2023.2 Bobcat
This proposed a release for oslo.rootwrap to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 7.0.1..0660a66
0660a66 Bump bandit
f02bdb9 Revert "Moves supported python runtimes from version 3.8 to 3.10"
90ea8f3 Moves supported python runtimes from version 3.8 to 3.10
3702177 Update master for stable/2023.1

Change-Id: If881591bca169d6cb4b0d30cd831d49958f99280
Signed-off-by: Elod Illes <elod.illes@est.tech>