Release final oslo.serialization for 2023.2 Bobcat
This proposed a release for oslo.serialization to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 5.1.1..a0ba2d7
a0ba2d7 Remove extra spaces in tox.ini
61c460f Fix compatibility with Python 3.8
b111b03 Revert "Moves supported python runtimes from version 3.8 to 3.10"
9ffd4e5 Moves supported python runtimes from version 3.8 to 3.10
a326ec5 Implement zoneinfo support to drop dependency to pytz
e3307fd Use the new openstack-python3-jobs template
864aa2f Update master for stable/2023.1

Change-Id: I3cef324df477af122ab754974063bf4fe17cd2af
Signed-off-by: Elod Illes <elod.illes@est.tech>