Release final oslo.service for 2023.2 Bobcat
This proposed a release for oslo.service to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 3.1.1..e94d47a
e94d47a Imported Translations from Zanata
4bff0d2 Bump bandit
b4c4869 Revert "Moves supported python runtimes from version 3.8 to 3.10"
a932419 Moves supported python runtimes from version 3.8 to 3.10
4341377 Update master for stable/2023.1

Change-Id: I9045130ed31761eb837c0c66133a6cf9f5c31d21
Signed-off-by: Elod Illes <elod.illes@est.tech>