Release final oslo.upgradecheck for 2023.2 Bobcat
This proposed a release for oslo.upgradecheck to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 2.1.1..cbee52e
cbee52e Revert "Moves supported python runtimes from version 3.8 to 3.10"
d84fbc1 Moves supported python runtimes from version 3.8 to 3.10
a8f3880 Update master for stable/2023.1

Change-Id: I3ba23a2210ba764ba32bcd93a999716d7a1127ca
Signed-off-by: Elod Illes <elod.illes@est.tech>