Release final oslo.utils for 2023.2 Bobcat
This proposed a release for oslo.utils to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 6.2.0..a5941e8
a5941e8 Imported Translations from Zanata

Change-Id: Ie7fc38b914d4207187e103967ac9371854f6db67
Signed-off-by: Elod Illes <elod.illes@est.tech>