Release final oslo.versionedobjects for 2023.2 Bobcat
This proposed a release for oslo.versionedobjects to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 3.1.0..6478669
f78601f Bump bandit and make oslo.versionedobjects compatible with latest rules
efe0cdc Imported Translations from Zanata
120e23d Revert "Moves supported python runtimes from version 3.8 to 3.10"
20d24cb Moves supported python runtimes from version 3.8 to 3.10
638a788 Update master for stable/2023.1
fb5a423 Imported Translations from Zanata

Change-Id: Ic312840b98fe818de4d21384303b5dda5b21d608
Signed-off-by: Elod Illes <elod.illes@est.tech>