Release final oslo.vmware for 2023.2 Bobcat
This proposed a release for oslo.vmware to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 4.1.1..4f97245
4f58211 Bump bandit and make oslo.messaging compatible with latest rules
abd720b Imported Translations from Zanata
f442d81 Revert "Moves supported python runtimes from version 3.8 to 3.10"
d6cdf97 Moves supported python runtimes from version 3.8 to 3.10
9451f96 Update master for stable/2023.1

Change-Id: If63afe5c5e5b77e5dd50ae9b4253d86c0d314f0c
Signed-off-by: Elod Illes <elod.illes@est.tech>