Release final ovsdbapp for 2023.2 Bobcat
This proposed a release for ovsdbapp to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 2.4.0..6f7aed7
6f7aed7 Add some basic usage documentation

Change-Id: I71c9fbfa0f070776ef9d1a8ec3f85cc3295352f1
Signed-off-by: Elod Illes <elod.illes@est.tech>