Release final sushy for 2023.2 Bobcat
This proposed a release for sushy to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 4.5.0..7c2a251
7c2a251 Requests must always have a read/connect timeout

Change-Id: I10c57ef9822a2ea962bfb741ca086472177645de
Signed-off-by: Elod Illes <elod.illes@est.tech>