Release final glance_store for 2023.2 Bobcat
This proposed a release for glance_store to get any final merged
changes out prior to the lib freeze for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, August 25th.

$ git log --oneline --no-merges 4.6.0..0c60291
0c60291 Imported Translations from Zanata
5a81f77 Imported Translations from Zanata

Change-Id: I47f2869de1bd238bed1deb13f7586b596a7b29cb