Release final osc-placement for 2023.2 Bobcat
This is a release for osc-placement to get any merged but
unreleased changes out prior to the client lib freeze for 2023.2 Bobcat
going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, September 1st.

$ git log --oneline --no-merges 4.1.0..ffa3822
ffa3822 Update master for stable/2023.1

Change-Id: If3d6894a0ddf168fbdc645870c347777c27062b4
Signed-off-by: Hervé Beraud <hberaud@redhat.com>