Release final python-barbicanclient for 2023.2 Bobcat
This is a release for python-barbicanclient to get any merged but
unreleased changes out prior to the client lib freeze for 2023.2 Bobcat
going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, September 1st.

$ git log --oneline --no-merges 5.5.0..7d9c28a
7d9c28a Fixed typo in the release notes

Change-Id: I0baff4f9d1f26aa27c677268366bcda2bcc08d0d
Signed-off-by: Hervé Beraud <hberaud@redhat.com>