Release final python-brick-cinderclient-ext for 2023.2 Bobcat
This is a release for python-brick-cinderclient-ext to get any merged but
unreleased changes out prior to the client lib freeze for 2023.2 Bobcat
going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, September 1st.

$ git log --oneline --no-merges 2.1.0..a0aa649
a0aa649 Update functional jobs for 2023.2
e0a4a0c Add Python 3.10 to setup.cfg metadata
1a33b25 Update master for stable/2023.1

Change-Id: I30d4a5f605b515d9e03e7cf73344496622b3a224
Signed-off-by: Hervé Beraud <hberaud@redhat.com>