Release final python-cloudkittyclient for 2023.2 Bobcat
This is a release for python-cloudkittyclient to get any merged but
unreleased changes out prior to the client lib freeze for 2023.2 Bobcat
going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, September 1st.

$ git log --oneline --no-merges 4.7.0..0835706
0835706 Update master for stable/2023.1

Change-Id: I80a23149c189246141b140ffd52ecf05c7124cbf
Signed-off-by: Hervé Beraud <hberaud@redhat.com>