Release final python-cyborgclient for 2023.2 Bobcat
This is a release for python-cyborgclient to get any merged but
unreleased changes out prior to the client lib freeze for 2023.2 Bobcat
going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, September 1st.

$ git log --oneline --no-merges 2.2.0..51c06d5
51c06d5 add nagative unittest for program deployable in cyborg module
dfff421 add negative unittest for deployable program in cyborg module
0d04902 add unittest for program deployable in cyborg module
f397f0d add unittest for delete accelerator with instance uuid
4a99d71 add unittest for show device profile with not exist id
f07398d add unittest for show device profile with id
f4331b8 line too long

Change-Id: I9c480456d3cf3891319cad859adb618ce9d5b6d9
Signed-off-by: Hervé Beraud <hberaud@redhat.com>