Release final python-designateclient for 2023.2 Bobcat
This is a release for python-designateclient to get any merged but
unreleased changes out prior to the client lib freeze for 2023.2 Bobcat
going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, September 1st.

$ git log --oneline --no-merges 5.2.0..4796812
4796812 Add python3.9 support in testing runtime
d7bca35 Update master for stable/2023.1
34292ee Add hacking check for line continuation backslash
a692fd5 Drop explicit queue configuration

Change-Id: I271f3ecf1c60a0405289b72d3dff97c1813c0eaf
Signed-off-by: Hervé Beraud <hberaud@redhat.com>