Release final python-ironic-inspector-client for 2023.2 Bobcat
This is a release for python-ironic-inspector-client to get any merged but
unreleased changes out prior to the client lib freeze for 2023.2 Bobcat
going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, September 1st.

Change-Id: Ie4ad16079fc4cf8b52e558c5009d52a4bd29f3c7
Signed-off-by: Hervé Beraud <hberaud@redhat.com>