Release final python-manilaclient for 2023.2 Bobcat
This is a release for python-manilaclient to get any merged but
unreleased changes out prior to the client lib freeze for 2023.2 Bobcat
going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, September 1st.

$ git log --oneline 4.5.1..2cdbf4e
2cdbf4e Merge "Deprecate "manila" CLI client"
8a3f65b Merge "Implement share backup"
32f1a9a Bump tox version
5e24577 Implement share backup
4eccd68 Allow restricting access rules
1734b45 Support for resource locks
ba9a388 Deprecate "manila" CLI client
7a15a2a Add '--count' option for snapshot list API.
0a0d610 [CI] switch to using osc for bootstrap

Change-Id: I67e146bd5a72dc529fb19297b592a636807e812f
Signed-off-by: Hervé Beraud <hberaud@redhat.com>