Release final python-masakariclient for 2023.2 Bobcat
This is a release for python-masakariclient to get any merged but
unreleased changes out prior to the client lib freeze for 2023.2 Bobcat
going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, September 1st.

$ git log --oneline --no-merges 8.1.0..8a832d9
8a832d9 Update master for stable/2023.1

Change-Id: Iae90d77ad8dd3baf0a53d14350c036dfbe8ca1f4
Signed-off-by: Hervé Beraud <hberaud@redhat.com>