Release final python-novaclient for 2023.2 Bobcat
This is a release for python-novaclient to get any merged but
unreleased changes out prior to the client lib freeze for 2023.2 Bobcat
going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, September 1st.

$ git log --oneline --no-merges 18.3.0..e464940f
e464940f Typo - nova CLI deprecation warning
67caa104 Update master for stable/2023.1

Change-Id: I1e3d71119f8dd4c689203d1c9e1d2b7c70f36f37
Signed-off-by: Hervé Beraud <hberaud@redhat.com>