Release final python-octaviaclient for 2023.2 Bobcat
This is a release for python-octaviaclient to get any merged but
unreleased changes out prior to the client lib freeze for 2023.2 Bobcat
going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, September 1st.

$ git log --oneline --no-merges 3.4.0..7ce6349
7ce6349 Fix validation of HSTS attributes
138eda3 Fix inconsistency in validation error message for HSTS
468fba0 Add extra validation for HSTS related arguments
15a8d74 Add support for HTTP Strict Transport Security
98beca9 Update master for stable/2023.1

Change-Id: I50f7c238140e2c1cabee13d1aa2e8868cd7907f5
Signed-off-by: Hervé Beraud <hberaud@redhat.com>