Release final python-saharaclient for 2023.2 Bobcat
This is a release for python-saharaclient to get any merged but
unreleased changes out prior to the client lib freeze for 2023.2 Bobcat
going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, September 1st.

$ git log --oneline --no-merges 4.1.0..693a0b6
131659a Update various links in the documentation
64f32c3 Fix deprecated zuul queue syntax
e455855 setup.cfg: Replace dashes with underscores

Change-Id: Ida703c847e9d951600dd5214e76f3f05ede51bb2
Signed-off-by: Hervé Beraud <hberaud@redhat.com>