Release final python-senlinclient for 2023.2 Bobcat
This is a release for python-senlinclient to get any merged but
unreleased changes out prior to the client lib freeze for 2023.2 Bobcat
going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, September 1st.

$ git log --oneline --no-merges 3.0.0..1644b1b
1644b1b Imported Translations from Zanata
9bf0bd6 Update master for stable/2023.1

Change-Id: I84c96ccda68982e01462125c021849c6be8f880a
Signed-off-by: Hervé Beraud <hberaud@redhat.com>