Release final python-troveclient for 2023.2 Bobcat
This is a release for python-troveclient to get any merged but
unreleased changes out prior to the client lib freeze for 2023.2 Bobcat
going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, September 1st.

$ git log --oneline --no-merges 8.2.0..0415e4b
0415e4b Drop openstack-lower-constraints-jobs

Change-Id: I5364282e22c3b1f4267664e525248f2256db7a96
Signed-off-by: Hervé Beraud <hberaud@redhat.com>