Release final python-vitrageclient for 2023.2 Bobcat
This is a release for python-vitrageclient to get any merged but
unreleased changes out prior to the client lib freeze for 2023.2 Bobcat
going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, September 1st.

$ git log --oneline --no-merges 4.7.0..cf150d2
cf150d2 Update master for stable/2023.1

Change-Id: I55c38f2113b1dbaddca308740eadc0221d579e7e
Signed-off-by: Hervé Beraud <hberaud@redhat.com>