[karbor] Transition Stein to End of Life
As announced on mailing list [1], this patch transitions the Stein
branch to End of Life. The last patch of the branch will be tagged with
stein-eol tag. stable/stein branch cannot be used anymore and will be
deleted if this patch merges.

This is needed as stable/stein is not actively maintained in recent
period, thus gates are mostly broken due to job failures. Besides,
by removing these branches, infra resources will be freed up, too.

Please try to identify any zuul job, that is defined outside of the
repositories in this patch (for example in openstack-zuul-jobs, etc.)
and won't be used anymore if stable/stein is deleted. Propose a job
removal patch for them.

Please +1 if the team is ready for us to proceed with this transition,
or -1 if there are still some activity on the branch and the team wants
to continue to maintain it.

[1] https://lists.openstack.org/pipermail/openstack-discuss/2023-September/034965.html

Change-Id: I8546722f6af1da0eb90306992a4fa2fec4f3714b