Release ansible-role-thales-hsm RC1 for 2023.2 Bobcat
This proposed a RC1 release for ansible-role-thales-hsm to get any merged changes
out prior to the RC1 deadline for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by early Friday, September 15.

Change-Id: I30ce104307680a5aeee70179b8d3c04c3e97fc97
Signed-off-by: Elod Illes <elod.illes@est.tech>