Release murano-dashboard RC1 for 2023.2 Bobcat
This proposed a RC1 release for murano-dashboard to get any merged changes
out prior to the RC1 deadline for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by early Friday, September 15.

Change-Id: Ia752901f778e16854bc7d544776b2ed217277c59
Signed-off-by: Elod Illes <elod.illes@est.tech>