Release networking-bagpipe RC1 for 2023.2 Bobcat
This proposed a RC1 release for networking-bagpipe to get any merged changes
out prior to the RC1 deadline for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by early Friday, September 15.

Change-Id: Ib611101b884954c3d71a1b6fbbf2e37e6d312dcc
Signed-off-by: Elod Illes <elod.illes@est.tech>