Release tacker-horizon RC1 for 2023.2 Bobcat
This proposed a RC1 release for tacker-horizon to get any merged changes
out prior to the RC1 deadline for 2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by early Friday, September 15.

Change-Id: I5879c8a83729f1592ec2a1e30736a0626f779515
Signed-off-by: Elod Illes <elod.illes@est.tech>