[sahara] Create 2023.2 branch for client and non-client libs
This creates the stable/2023.2 branch for all client and non-client
libraries now that we are past the library freeze.

Note that branching deadline was last week, so please respond as soon
as possible with a +1 (or -1 if for example a gate fixing patch is on
the way).

This patch will be merged, if there is no response from the team before
Friday.

Change-Id: I78bc67fcf245d9b1cdd3753e97560bceb30e70bb