Release ec2api-tempest-plugin for 2023.2 Bobcat
This patch proposes a release for ec2api-tempest-plugin
to get the latest merged changes out prior to the final
2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by early Friday, September 22.

Change-Id: I6c89c3e4e6aac3ac1e74719cf2bceebdb90869dc
Signed-off-by: Elod Illes <elod.illes@est.tech>