Release trove-tempest-plugin for 2023.2 Bobcat
This patch proposes a release for trove-tempest-plugin
to get the latest merged changes out prior to the final
2023.2 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by early Friday, September 22.

Change-Id: I17182d6e54e46d3681b29f9e5f5cad52dd230019
Signed-off-by: Elod Illes <elod.illes@est.tech>