Final RC patch for ansible-role-thales-hsm
This creates a new RC before the final release candidate deadline to
release the latest fixes and/or translations.

Please ack with a +1, or if there is some reason to hold off, leave a -1.

Patches with no response will be merged by Friday, Sept 29th.

$ git log --oneline --no-merges 6.0.0.0rc1..f74caaa
f74caaa This is a combination of 2 commits.

Change-Id: Id4fa6da2c1236aab39d718ed016cc0274bf58359
Signed-off-by: Hervé Beraud <hberaud@redhat.com>