Final RC patch for zun
This creates a new RC before the final release candidate deadline to
release the latest fixes and/or translations.

Please ack with a +1, or if there is some reason to hold off, leave a -1.

Patches with no response will be merged by Friday, Sept 29th.

$ git log --oneline --no-merges 12.0.0.0rc1..9c240d27
9c240d27 Temporarily disable multinode job
b368b6fc Update TOX_CONSTRAINTS_FILE for stable/2023.2
1b0c68be Update .gitreview for stable/2023.2

Change-Id: Id933fc0d0e1b6411e789f2b8ad63377c282513b0
Signed-off-by: Hervé Beraud <hberaud@redhat.com>