Fix approval check for distributed leadership type
Teams with distributed leadership type have release liaisons listed
in governance repository's projects.yaml file instead of a single PTL.
This was introduced in the check_approval.py script, but it did not
work (though we did not realise this as usually releases repository's
release_liaisons.yaml also contains the release liaisons). This patch
fixes the check for distributed leadership type.

Change-Id: I93fbf2be0f0da85dedfe9b9f8d5876ca5b9e4b2e