Change blazar-nova's type from library to other
Some cycles ago blazar-nova's release model changed from
cycle-with-intermediary to cycle-with-rc, as this deliverable is highly
dependent on nova and usually needs to wait for nova's release.

When this change happened blazar-nova's release type remained library,
which causes that during the cycle the release process guidelines and
tooling always shows that blazar-nova should have a release at library
freeze deadline, which is exactly why the release model change was
intended to avoid. So this patch changes the release type from library
to 'other' so that release tooling won't give false alarms about
missing releases at library freeze time.

Change-Id: If8f2e682df9e4432704f48ee26b659f87ebcc1a6