fix tooling to retrieve unbranched deliverables based release id
list unbranched deliverables is not designed to work with new release
id.

These changes add a command to allow to retrieve series id from a shell
script by passing the name of the series.

This new command is called by the list unbranched deliverables to check
if the corresponding stable branch exists (based on the release id).

Change-Id: I102b90c5221e57dd5697e2e81aa7edb8615ce445