release cliff for caracal-1 milestone
This is a library release for cliff for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 4.3.0..3ee0725
606644c Handle complex objects in yaml formatter better
2f3e17d Add pre-commit
13424d1 Fix pre-commit issues
e02f89c Autofit table output if stdout is a tty
6096869 Fix flake8 violation E721
989c440 Use upper-constraint in doc generation
f2215ba Clarification of the algorithm used

Change-Id: Ie18f21627ce3cf4bc06b386af7ed7622558cbf3a
Signed-off-by: Hervé Beraud <hberaud@redhat.com>