release keystoneauth for caracal-1 milestone
This is a library release for keystoneauth for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 5.3.0..44e477d
74f0524 Drop parameters when connecting to a redirected endpoint
efba3f9 Update master for stable/2023.2
c800dce Keep request-id at redirect
66cd04e Add project URLs from README as packaging metadata

Change-Id: I4d7998656316f046d54e49521b6a2a4b86bb68af
Signed-off-by: Hervé Beraud <hberaud@redhat.com>