release keystonemiddleware for caracal-1 milestone
This is a library release for keystonemiddleware for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 10.4.1..9ade074
53b4cb2 External OAuth2.0 Authorization Server Support
de15a61 External OAuth2.0 Authorization Server Support
7033768 auth_token: fix issue when data in cache gets corrupted

Change-Id: Iaa5381e2398150cab56d1c8b0a3673d84e909197
Signed-off-by: Hervé Beraud <hberaud@redhat.com>