release metalsmith for caracal-1 milestone
This is a library release for metalsmith for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 2.0.1..49c8cab
66a94ef Document LP as official bugtracker
e4fd02f Metalsmith in maintenance mode
71192b1 Update master for stable/2023.2
9c928d3 Set role default metalsmith_provisioning_timeout

Change-Id: I6760fc72d1cd8f1acbeb4f0837f4caa65b8eaf4a
Signed-off-by: Hervé Beraud <hberaud@redhat.com>