release mistral-extra for caracal-1 milestone
This is a library release for mistral-extra for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 13.1.0..2e2cf63
3a86bd3 fix BaremetalIntrospectionAction loading failed
6482a86 Rename trove instances.edit to instances.update

Change-Id: I87b41dd0fcb597e85cdf1a51597fdff5a69d76df
Signed-off-by: Hervé Beraud <hberaud@redhat.com>