release neutron-lib for caracal-1 milestone
This is a library release for neutron-lib for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 3.8.0..64c693721
8bb438dd Switch to py311 sqlalchemy master job
343b2db5 Remove unmaintained API report tooling
825d292c (bug/2043141) Change test_sqlalchemytypes.py to support sqlalchemy 2.0
b7598dab vpnaas: add support for more ciphers (auth, encryption, pfs modes)
08b8c6d3 Add the "cancellable" flag to the ``CallbacksManager`` events
1da71876 Imported Translations from Zanata
bdebe1de Update master for stable/2023.2
7de2cb33 api-ref: Move sfc api-ref to neutron-lib

Change-Id: Ied01bd54f8ed847e06c06d8062f6bb7e751d5bf5
Signed-off-by: Hervé Beraud <hberaud@redhat.com>