release os-brick for caracal-1 milestone
This is a library release for os-brick for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 6.4.0..fb23a40
bbc732c Revert "Dell PowerFlex password appears in plain text when creating a volume"
33661ec Dell PowerFlex password appears in plain text when creating a volume from an image
42d9a89 Dell Powerflex: Add new VOLUME_ALREADY_MAPPED_ERROR
034a6f2 Update master for stable/2023.2
eea10de Use generic testing template

Change-Id: Ifaaf533f1f9c76b81714eeefe53bb8784e13d111
Signed-off-by: Hervé Beraud <hberaud@redhat.com>