release os-vif for caracal-1 milestone
This is a library release for os-vif for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 3.2.0..458e8d5
458e8d5 Update master for stable/2023.2
0a0dec3 Don't break traffic if port already exists

Change-Id: Ic57072c43661f89486bbb8dbcee24938594fc70c
Signed-off-by: Hervé Beraud <hberaud@redhat.com>