release osc-lib for caracal-1 milestone
This is a library release for osc-lib for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 2.8.1..3f47018
32b640f Update master for stable/2023.2
9af8b45 utils: Don't warn for partial formatter classes

Change-Id: Ib09d6178a13e66fe0cb78454b6afedf1184ec9ba
Signed-off-by: Hervé Beraud <hberaud@redhat.com>