release oslo.context for caracal-1 milestone
This is a library release for oslo.context for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 5.2.0..8d25cad
059f9ba Add method for getting redacted copy of context
1e95f85 Update master for stable/2023.2
e70a618 Bump bandit

Change-Id: Id514a01f1a03723e2cb67bb69befe29614023fa3
Signed-off-by: Hervé Beraud <hberaud@redhat.com>