release oslo.i18n for caracal-1 milestone
This is a library release for oslo.i18n for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 6.1.0..4bcda78
3dd3fa7 Update master for stable/2023.2
05f9961 Bump bandit
77a994a The locale.getdefaultlocale is deprecated

Change-Id: I83182970614ab32a93dbcb4babec5ffde7fdbcea
Signed-off-by: Hervé Beraud <hberaud@redhat.com>