release oslo.log for caracal-1 milestone
This is a library release for oslo.log for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 5.3.0..b5b8c30
f3693db Deprecate Windows support
ca1ef3b Update master for stable/2023.2
390a1d1 Catch RuntimeError when loading log config file

Change-Id: I4c02f0078c3056296bbdf2cd6b0c70e9a5f6c5d0
Signed-off-by: Hervé Beraud <hberaud@redhat.com>