release oslo.vmware for caracal-1 milestone
This is a library release for oslo.vmware for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 4.2.0..60da85c
60da85c Imported Translations from Zanata
19b4d3c Add PBM WSDL for vSphere 8.0
da01521 Update master for stable/2023.2

Change-Id: Ib3ed1763e0415bf39edc187eed2cb62c694b2f2f
Signed-off-by: Hervé Beraud <hberaud@redhat.com>