release ovsdbapp for caracal-1 milestone
This is a library release for ovsdbapp for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 2.4.1..8e55b0b7
ef564d6 nb: allow to add route even if learned route exists
02a7b2b TAAS: Add commands for creating Mirrors
dae455b nb: provide 'route_table' in lr-route-{add,del,list}
033b6e3 nb: add 'nexthop' argument to 'lr_route_del'
6193545 Handle events with conditions and match_fn
dd56fd5 venv: unit tests don't work for 'hardware_vtep'
a39b4f8 Update master for stable/2023.2

Change-Id: I411ad6c47b3e5c63ef2762934d19d00483eca83e
Signed-off-by: Hervé Beraud <hberaud@redhat.com>