release python-cyborgclient for caracal-1 milestone
This is a library release for python-cyborgclient for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 2.2.1..f2d73d3
f2d73d3 Update master for stable/2023.2
528144a Add client api for enable and disable device
2ca2978 resolve keyerror for string format
0b8f60c add ut for delete attribute client exception error
de34fc8 Update master for stable/2023.1
8c19a8f refactor the getid method base.py

Change-Id: I742c77a4f7f4d9df5a32c0be882354bcd088bbd6
Signed-off-by: Hervé Beraud <hberaud@redhat.com>