release python-magnumclient for caracal-1 milestone
This is a library release for python-magnumclient for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 4.2.0..aa0d180
fb28368 Bugfix: master_lb_enabled flag now works as intended.
3d5ab01 setup.cfg: rename deprecated dash-separated options
7f00f41 Update master for stable/2023.2
31a6c8f Use TOX_CONSTRAINTS_FILE
bb49cef Add python 3.10 to setup.cfg metadata
13cafc1 Clean up tox.ini
c30f10f Remove translation sections from setup.cfg

Change-Id: Ie02ecf629db05ca7b15aa70a34a556024341d186
Signed-off-by: Hervé Beraud <hberaud@redhat.com>