release python-manilaclient for caracal-1 milestone
This is a library release for python-manilaclient for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 4.6.0..64f5814
a870c13 Fix default share type resolution in OSC
244e87a Fix "test_share_access_show" osc functional test
946c113 Fix decoder path and add transfer commands
a078139 Update master for stable/2023.2

Change-Id: I29c377f7164501508a71aa78385ce7356d41da0d
Signed-off-by: Hervé Beraud <hberaud@redhat.com>