release python-masakariclient for caracal-1 milestone
This is a library release for python-masakariclient for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 8.2.0..b43a4ce
b43a4ce support for Masakari VMove

Change-Id: I194769ce10436d2eca63e01b7d1ebfd3eece99b0
Signed-off-by: Hervé Beraud <hberaud@redhat.com>