release python-neutronclient for caracal-1 milestone
This is a library release for python-neutronclient for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 11.0.0..8ae44f8
ee6f6a1 Drop unused simplejson
8c983fa Imported Translations from Zanata
b9152a5 OSC: Remove SFC V2 calls to neutronclient
396432a OSC: Remove VPNAAS V2 calls to neutronclient
6597c42 Update master for stable/2023.2
fccfe8c OSC: Remove FWAAS V2 calls to neutronclient

Change-Id: Ib81a603d49e43f9bbc28007b251375821acb65be
Signed-off-by: Hervé Beraud <hberaud@redhat.com>