release python-octaviaclient for caracal-1 milestone
This is a library release for python-octaviaclient for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 3.5.0..db6b62e
db6b62e Set default value for HSTS options to None

Change-Id: I4f9163e8721726259d461e29c71b0d2d8522abd6
Signed-off-by: Hervé Beraud <hberaud@redhat.com>