release python-venusclient for caracal-1 milestone
This is a library release for python-venusclient for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 2.1.0..9208ade
9208ade Add delete-anomaly-rule in venusclient CLI
956a08e Combine 4 get-type-xxx to get-search-params in venusclient CLI
73f507b Add 'get-anomaly-rule' to venusclient
5a2b1f5 Add parameter desc of analyse-log of venusclient CLI
38abcf7 Add add_anomaly_rule to venusclient
594f204 modify search logs params in venusclient
8d6b6f6 Add set-custom-config subcommand to venusclient
1ca4f68 Document Format Modification
5e791c3 Update master for stable/2023.2
243f335 Adding get-instance-callchain subcommand to venusclient

Change-Id: I13787c74cfe2c7eb8298022b3a7b243c5f143612
Signed-off-by: Hervé Beraud <hberaud@redhat.com>