release sushy for caracal-1 milestone
This is a library release for sushy for the caracal-1
milestone. This repo includes commits that have not been releases for
this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

$ git log --oneline --no-merges 4.5.1..b87bc5
4ff512a Fix missing ETag when patching Redfish resource
82772d3 Update master for stable/2023.2
89a0988 Fix wrong _get_registry logic in ResourceBase

Change-Id: I334b563b27e2788507ddf8b24bde2deebdb68ea3
Signed-off-by: Hervé Beraud <hberaud@redhat.com>