[cloudkitty] Transition Ussuri to End of Life
With the new resolution '2023-07-24 Unmaintained status replaces
Extended Maintenance' [1] the last 3 EM branches (Xena, Wallaby,
Victoria) will transition to Unmaintained, but older branches (Ussuri,
Train) should move to End of Life unless someone steps up as maintainer
of this branch for this project.

To signal this, please add a -1 for this patch if you or the team want
to keep this branch open for further maintenance.
If there is no response from anyone within a reasonable time frame,
then PTLs/release liaisons will be asked to confirm the EOL transition
with their +1.

[1] https://governance.openstack.org/tc/resolutions/20230724-unmaintained-branches.html

Change-Id: I5c90d0d50d11cd6e09e33fb2524dfc3e0199d4a7