Caracal-2 release for oslo.utils
This is the Caracal-2 milestone release for oslo.utils.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If the team is happy with the current release hash, please +1 this
patch and we will process the release. If you need to get some
important changes merged first, leave a -1 and update the patch with
a new commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed after January 11th.

Change-Id: Iccc767cee67fdf090c060ab7c9bd68942ecc5bc1
Signed-off-by: Elod Illes <elod.illes@est.tech>