Release horizon 23.4.0(Caracal)
This is the first release for Caracal. The minor version is bumped as we
have changes in requirements.txt(Django>=3.2,<3.3 changed to >=3.2,<4.3).
Horizon bumps the upper version of Django to Django4.2. So this patch
cut a new release of the horizon so that all plugins can consume
this change. For more information please see [1]

[1] https://review.opendev.org/c/openstack/horizon/+/905218

Change-Id: Ibfb863c72bd07d404ded8ed03503e4bb097a6cc1