Release OpenStack-Ansible 2023.2 (Bobcat)
Despite having changes for new minor release, we historically do first
minor release for series when we sure that it is safe to perform upgrades
to this release series.

With that, we still want to mark it as a "bugfix" release as we are not
completely
sure that we can recommend upgrades at this point.

Change-Id: Iacb14bcd782745e2ecdbe8749c57430b957b460c