Release ansible-role-lunasa-hsm for stable/zed
This release picks up new commits to ansible-role-lunasa-hsm since
the last release from stable/zed.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 6.0.0..d93586c
d93586c Fix vtl verification
bbf8f85 Fix permissions error during client installation
44b5477 Update .gitreview for stable/zed

Change-Id: I47f71b03b29f68cdcf23e90ccd40785eb4e2b1d2
Signed-off-by: Elod Illes <elod.illes@est.tech>