Release cloudkitty for stable/zed
This release picks up new commits to cloudkitty since
the last release from stable/zed.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 17.0.0..56ec4f1
56ec4f1 Fix random unit test failures
d848043 Fix retrieval of reprocessing tasks
44a7db9 Fix PyScripts processing
12a84c5 [docs] Install cloudkitty requirements using constraints
64b92f3 Update TOX_CONSTRAINTS_FILE for stable/zed
e146972 Update .gitreview for stable/zed

Change-Id: I73ed91ac817fc7951918d21a48ebb7133cfac0b7
Signed-off-by: Elod Illes <elod.illes@est.tech>