Release ironic-inspector for stable/zed
This release picks up new commits to ironic-inspector since
the last release from stable/zed.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 11.1.0..9ea9f61
9ea9f61 Handle bracketed IPv6 redfish_address
76e277b [zed-only] Fix functional tests
e1932cd Make reno ignore bugfix eol tags
6c3cbd1 [stable-only] Fixes for tox 4.0 compatability
c2e4d64 Update TOX_CONSTRAINTS_FILE for stable/zed
a2caaab Update .gitreview for stable/zed

Change-Id: I3c1dbf135c099276d6d413f4279e4074e36623e3
Signed-off-by: Elod Illes <elod.illes@est.tech>