Release ironic-python-agent-builder for stable/zed
This release picks up new commits to ironic-python-agent-builder since
the last release from stable/zed.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 5.0.0..f560170
f560170 Drop TripleO job
e2c1c82 Extend the DIB_CHECKSUM variable usage
dbd6634 Add checksum generation support
b097986 Add libaio engine for fio on Centos/Rhel9
cb5efad Remove /var/tmp/* from images
daab68f Update TOX_CONSTRAINTS_FILE for stable/zed
f6d15d0 Update .gitreview for stable/zed

Change-Id: I920fde26c3f86c67b031a06d8738982ddf69e0b4
Signed-off-by: Elod Illes <elod.illes@est.tech>