Release keystone for stable/zed
This release picks up new commits to keystone since
the last release from stable/zed.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 22.0.1..72a4fc0f3
72a4fc0f3 Add an option to randomize LDAP urls list

Change-Id: I6b6e996d3801073496f276fa044f9dc9389c4768
Signed-off-by: Elod Illes <elod.illes@est.tech>