Release keystoneauth for stable/zed
This release picks up new commits to keystoneauth since
the last release from stable/zed.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 5.0.0..61c3dc7
61c3dc7 Make v3oidcpassword send client_id
1b262cd Update TOX_CONSTRAINTS_FILE for stable/zed
7053ec9 Update .gitreview for stable/zed

Change-Id: I715ae59f2a9c77c3339865e2de106f8fc6327830
Signed-off-by: Elod Illes <elod.illes@est.tech>