Release keystonemiddleware for stable/zed
This release picks up new commits to keystonemiddleware since
the last release from stable/zed.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 10.1.0..b61d312
b61d312 Add timeout for requests
f602fa7 Remove cache invalidation when using expired token
46778cd Update TOX_CONSTRAINTS_FILE for stable/zed
0d3dd74 Update .gitreview for stable/zed

Change-Id: Iccc6a12ac8e3e609e10ad4fdee2bb59dbed8c965
Signed-off-by: Elod Illes <elod.illes@est.tech>