Release metalsmith for stable/zed
This release picks up new commits to metalsmith since
the last release from stable/zed.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 1.8.0..47f5953
47f5953 Stable-Only: CI: Force install an appropriate openstacsdk version
176f48e stable-only: Constrain the upper Ansible version
cc388a4 test_provision: don't assert provision_node call order
392593d list_instances - cache allocations
ea3a920 Get ports by 'binding:host_id' query filter
b5f695f Use a network cache in Instance
0e839b0 Fixes for tox 4.0
f586ede Allow to use Ansible OpenStack Col. 1.x.x with openstacksdk >=0.99.0
008425a Update TOX_CONSTRAINTS_FILE for stable/zed
8e7fba3 Update .gitreview for stable/zed

Change-Id: Ic8eedc47f6434451861e1c57f325d0b4f4c3ebb4
Signed-off-by: Elod Illes <elod.illes@est.tech>