Release mistral for stable/zed
This release picks up new commits to mistral since
the last release from stable/zed.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 15.0.0..0958cb23
0958cb23 Fix sorting error when type is different
b450848a Update zuul queue configuration
041e9a33 Update TOX_CONSTRAINTS_FILE for stable/zed
c8b29101 Update .gitreview for stable/zed

Change-Id: I1b6296648adeec73b960dd7d23e93642d789d60e
Signed-off-by: Elod Illes <elod.illes@est.tech>