Release networking-generic-switch for stable/zed
This release picks up new commits to networking-generic-switch since
the last release from stable/zed.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 7.0.0..5ea181e
5ea181e Fix regression plugging 802.3ad port group
0f1c5dd Fixes for tox 4.0
cef630a Update TOX_CONSTRAINTS_FILE for stable/zed
248fafa Update .gitreview for stable/zed

Change-Id: I407f3c6fe7e77c75ff5260503e2181c462992409
Signed-off-by: Elod Illes <elod.illes@est.tech>