Release neutron-fwaas for stable/zed
This release picks up new commits to neutron-fwaas since
the last release from stable/zed.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 17.0.0..7c860a0c7
9cd7d7048 Add CONTEXT_WRITER to _get_default_fwg_id
2a4f85e9c Add get_firewall_groups_count method
1fc233a50 Use upper-constraints in docs job
7545e6a26 Fix tox.ini with tox4
f4ce1441a Update TOX_CONSTRAINTS_FILE for stable/zed
5ef0ce7c6 Update .gitreview for stable/zed

Change-Id: I38d7b20e3285f25c6cf34d3470a57de369b60f61
Signed-off-by: Elod Illes <elod.illes@est.tech>