Release nova for stable/zed
This release picks up new commits to nova since
the last release from stable/zed.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 26.2.1..226f3e95c1
e41962f5fa Do not untrack resources of a server being unshelved
7422642dd6 Reproduce bug #2025480 in a functional test

Change-Id: I49a68a7df02c793812150d12cec2b4decd07f89f
Signed-off-by: Elod Illes <elod.illes@est.tech>