Release octavia-lib for stable/zed
This release picks up new commits to octavia-lib since
the last release from stable/zed.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 3.1.0..3f764c5
3f764c5 Fix a possible receive timeout
ca1058c Update TOX_CONSTRAINTS_FILE for stable/zed
4e6cc14 Update .gitreview for stable/zed

Change-Id: I4351fbf228782347387096385bd2ea6e26d87fb7
Signed-off-by: Elod Illes <elod.illes@est.tech>