Release octavia for stable/zed
This release picks up new commits to octavia since
the last release from stable/zed.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 11.0.1..000b577f
df8ebcdb Don't update the management port when calculating delta
f7265825 Fix issue with certificates with no subject or CN
85f94c25 Stable-only: Cap hacking to < 6.1.0
07c33fec Add Noop Certificate Manager
d09065c4 Fix amphorae in ERROR during the failover
dd849b4c Reduce duration of failovers with amphora in ERROR
1dadaff5 Fix timeout duration in start_vrrp_service during failovers
27060603 Retry to set loadbalancer prov status on failures
4c97b585 Fix race condition in members batch update API call
752560a5 Fix incorrect removal of IP rules in the amphora
22d616c5 Fix octavia-status with amphorav2
5841efb7 Fix upgrade check not working
146b3fe4 Fix create_server_group in compute noop
106376d3 Fix text/plain mime type with healthcheck endpoint
68a3379c Fix amphorav1 member deletion bug
3a1e35da Fix persistence_granularity default value
c014bb7b Fix typo in lvs-masquerade.sh
b9f01c7d Fix job book deletion issue in taskflow
24024fb2 fix haproxy_count to display the number of running processes
68d464fa allowed_cidr validation for additional_vips
658173e5 Fix failover when the last listener is deleted

Change-Id: I2d0d3c334024b089813ab9f6989670b03103e29d
Signed-off-by: Elod Illes <elod.illes@est.tech>