Release os-vif for stable/zed
This release picks up new commits to os-vif since
the last release from stable/zed.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 3.0.0..51b7f48
51b7f48 Move mtu update request into ovsdb transaction
a602eca Update TOX_CONSTRAINTS_FILE for stable/zed
37fd20f Update .gitreview for stable/zed

Change-Id: Iefde105c62db4c899fc38b89345e120fdf1d6391
Signed-off-by: Elod Illes <elod.illes@est.tech>