Release python-openstackclient for stable/zed
This release picks up new commits to python-openstackclient since
the last release from stable/zed.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 6.0.0..0c07b165
0c07b165 [stable-only] CI: Drop all tips jobs
48148e7a Fix "access rule" commands to only use ID
fc80188e Fix pep issue in the network service provider
7d702baf compute: Add support for microversion 2.93
0118d57c compute: Require image when rebuilding a volume-backed server
d28c2584 Update TOX_CONSTRAINTS_FILE for stable/zed
2b19e97e Update .gitreview for stable/zed

Change-Id: I08de7eafa4db755bd02cbf51c418216dd112af10
Signed-off-by: Elod Illes <elod.illes@est.tech>