Release sushy for stable/zed
This release picks up new commits to sushy since
the last release from stable/zed.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 4.3.3..b11baf5
bb00d94 Handle exceptions after re-authentication
47e4e55 Remove version field from iLO error
8ff0d10 Requests must always have a read/connect timeout
b7d2456 Retry on ilo state error
4ee0a97 workaround: requests verify handling if env is set
e4108b3 Exclude all files starting with . from flake8 tests
3e16da9 Fix volume deletion on newer iDRACs
e39c25f Retry on iDRAC SYS518 errors for all requests
5e2807a Handle proper code_status in unit test
69b014c Handle a different error code for missing TransferProtocolType

Change-Id: Id7bd14336259904280b780b67f82b0c26c74cf50
Signed-off-by: Elod Illes <elod.illes@est.tech>