Release trove for stable/zed
This release picks up new commits to trove since
the last release from stable/zed.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 18.0.0..c8579f51
c8579f51 Fix guest agent failed to use swiftclient over TLS
e260d048 Add region name to trove-guestagent
9f8ed20e [CI]: Don't run periodic jobs on stable branchs
b195a6ab Fix "create database" failed for postgress
7c8b9053 Fixes permission problem when restoring backup
d99811a3 Change settings for stable/zed
392cf77d Update TOX_CONSTRAINTS_FILE for stable/zed
15965491 Update .gitreview for stable/zed

Change-Id: Ia34c56366d28772762d95d84abefc451f9f5abdb
Signed-off-by: Elod Illes <elod.illes@est.tech>