Release zun for stable/zed
This release picks up new commits to zun since
the last release from stable/zed.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 10.0.0..13379fd9
13379fd9 Specify a specific version of docker
d5f102d7 Fix circular import
5c2f205e Fix tox4 error
54c66fc3 Disable git owner check for zun
c3ddf2ec Handle io_service_bytes_recursive is None
1656d8ff Install zun from stable branch

Change-Id: I6cac04341b4fd816f8cd754e06063820b8d1badb
Signed-off-by: Elod Illes <elod.illes@est.tech>