[tempest] Tag yoga-last
Yoga branch is moving to unmaintained state now[1]. Tempest and
tempest plugins are branchless which means master version of
Tempest and its plugins is used to test the supported stable
branches.

Once any stable branch is moved to unmaintained (previously EM)
state then, Tempest and its plugins compatible tag[2] needs to be
released so that we can keep testing the unmaintained stable branches
with that tag once master Tempest and its plugins are not compatible[3].

Tagging 'yoga-last' with the same hash of new tag 37.0.0 which mark
the end of support for stable/yoga in Tempest.

[1] https://review.opendev.org/q/topic:%22yoga-unmaintained%22
[2] https://docs.openstack.org/tempest/latest/stable_branch_support_policy.html
[3] https://docs.openstack.org/tempest/latest/tempest_and_plugins_compatible_version_policy.html

Change-Id: I1ae5708e6040cb17c24f54abbad2f4c00b3bf3a3