[zun-tempest-plugin] Tag yoga-last
Yoga branch is in unmaintained state now[1]. Tempest and tempest plugins
are branchless which means master version of Tempest and its plugins is
used to test the supported stable branches.

Once any stable branch is moved to unmaintained (previously for EM) state
then, Tempest and its plugins compatible tag[2][3] needs to be released
so that we can keep testing the unmaintained stable branches with that tag.

Tagging 'yoga-last' with the latest hash which is nothing but 4.8.0 tag.

[1] https://review.opendev.org/q/topic:%22yoga-unmaintained%22
[2] https://docs.openstack.org/tempest/latest/stable_branch_support_policy.html
[3] https://docs.openstack.org/tempest/latest/tempest_and_plugins_compatible_version_policy.html

Change-Id: I7745163f7c14cec2ae9d626ccc2c2031513f329a