[keystone-tempest-plugin] Tag yoga-last
Yoga branch is in unmaintained state now[1]. Tempest and
tempest plugins are branchless which means master version
of Tempest and its plugins is used to test the supported
stable branches.

Once any stable branch is moved to unmaintained (previously
for EM) state then, Tempest and its plugins compatible tag[2][3]
needs to be released so that we can keep testing the unmaintained
stable branches with that tag.

Tagging 'yoga-last' with the already released 0.13.0 tag hash
(which is latest hash).

[1] https://review.opendev.org/q/topic:%22yoga-unmaintained%22
[2] https://docs.openstack.org/tempest/latest/stable_branch_support_policy.html
[3] https://docs.openstack.org/tempest/latest/tempest_and_plugins_compatible_version_policy.html

Change-Id: I34f1d29e88bcaf6c19a88d983b1744dd92eade1a