Allow <series>-last tag to be released for Unmaintained release
'<series>-last' tag is used to tag Tempest and its plugins to
work as last tag for EM release. Now with Unmaintained release
model, we need to allow this tag for Unmaintianed release also.
It will help Unmaintained branches to be tested with a compatible
version of Tempest and its plugin.

Change-Id: I7f58ad4570aead3b73bfab5c47cd29fa7a0a1700