Release final oslo.limit for 2024.1 Caracal
This proposed a release for oslo.limit to get any final merged
changes out prior to the lib freeze for 2024.1 Bobcat going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, February 22nd.

Change-Id: I0dab48fccec585d5b7c14e0f40cd059c37c2f31c
Signed-off-by: Elod Illes <elod.illes@est.tech>