Release final osc-placement for 2024.1 Caracal
This is a release for osc-placement to get any merged but
unreleased changes out prior to the client lib freeze for
2024.1 Caracal going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, February 29th.

Change-Id: If6ef01ad20916e1e8e70ff7088e1d526401d48b4
Signed-off-by: Elod Illes <elod.illes@est.tech>