Release final python-adjutantclient for 2024.1 Caracal
This is a release for python-adjutantclient to get any merged but
unreleased changes out prior to the client lib freeze for
2024.1 Caracal going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, February 29th.

Change-Id: I2efa4448e210bac59fcc1f46f8d2c5e979f04302
Signed-off-by: Elod Illes <elod.illes@est.tech>