Release final python-neutronclient for 2024.1 Caracal
This is a release for python-neutronclient to get any merged but
unreleased changes out prior to the client lib freeze for
2024.1 Caracal going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, February 29th.

Change-Id: I57029f311a7bad98b7ab4938b7c7de40823d4fe5
Signed-off-by: Elod Illes <elod.illes@est.tech>