Release final python-vitrageclient for 2024.1 Caracal
This is a release for python-vitrageclient to get any merged but
unreleased changes out prior to the client lib freeze for
2024.1 Caracal going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, February 29th.

Change-Id: I63246d2403a82aab8742304b4af3a096f4756109
Signed-off-by: Elod Illes <elod.illes@est.tech>