[ironic] Transition EM branches to Unmaintained
With the new TC resolution the community replaces Extended Maintenance
with Unmaintained status [1].

So this patch transitions victoria, wallaby, and xena EM branches to
Unmaintained as follows:
- tip of the given branch will be tagged with <series>-eom tag
- unmaintained/<series> branch will be cut from <series>-eom tag
- stable/<series> branch will be deleted

Some further notes:
- stable policy still applies
- backports can be pushed against unmaintained/<series> branch
- openstack-unmaintained-core group have the rights to merge patches

For further details please read the resolution [1] & review and +1 this
patch if this is correct as soon as possible.

This patch also officially EOLs two branches of sushy-cli which were
effectively unmaintained anyway and not available in releases after
wallaby.

[1] https://governance.openstack.org/tc/resolutions/20230724-unmaintained-branches.html

Change-Id: Ifa4e08ffbab548cd7e15bf88ae7201e92efeb7ac