[karbor] Transition EM branches to EOL
With the new TC resolution the community replaces Extended Maintenance
with Unmaintained status [1].

Since Karbor project is retired after victoria, it makes sense to move
its victoria branches to EOL instead of moving it to Unmaintained.
If anyone wants to resurrect the project then the work should be
started at the master branch instead.

[1] https://governance.openstack.org/tc/resolutions/20230724-unmaintained-branches.html

Change-Id: I4fee731a235546e36628a084cd5a3d0782752cfc