Create new oslo.context release for 2024.1
This creates a new release of oslo.context for 2024.1. This is required
to fix the heat unit tests failure which is currently blocking us from
bumping oslo.messaging version in u-c to the final 2024.1 release[1].

Because the change introduced a new interface which may be classified
as a new feature, a minor version bump is proposed, rather than bug
fix version bump.

[1] https://review.opendev.org/c/openstack/requirements/+/909906

Change-Id: Iabeaebe7ab0afa0eb4ba7828ab37342537baae2c