[barbican] Create 2024.1 branch for client and non-client libs
This creates the stable/2024.1 branch for all client and non-client
libraries now that we are past the library freeze.

Note that branching deadline was last week, so please respond as soon
as possible with a +1 (or -1 if for example a gate fixing patch is on
the way).

This patch will be merged, if there is no response from the team before
Friday.

Change-Id: If75ec267ff676da929024ec326f7f29df634345b