[murano] Create 2024.1 branch for client and non-client libs
This creates the stable/2024.1 branch for all client and non-client
libraries now that we are past the library freeze.

Note that branching deadline was last week, so please respond as soon
as possible with a +1 (or -1 if for example a gate fixing patch is on
the way).

This patch will be merged, if there is no response from the team before
Friday.

Change-Id: I7aaee99863edef94a3ea8d4e297e5f57cc2f2734