Drop murano deliverables from caracal
With a recent resolution[0] the Technical Committee has decided to
remove the murano project from the current release.

This patch removes the deliverables files for the affected repos where
no release has yet been made for this cycle.

The python-muranoclient repo has already been released, there is not
much we can do about that, for consistency the corresponding file is
left in place.

[0] https://governance.openstack.org/tc/resolutions/20240227-remove-murano-from-2024-1-release.html

Change-Id: Id097507c131cf8113e2a65542f273f1738e40519