Set Xena, Wallaby and Victoria state as Unmaintained
With the new TC resolution the community replaces Extended Maintenance
with Unmaintained status [1].

As most of the projects are ready with the transition [2], this patch
sets Xena, Wallaby and Victoria releases as unmaintained.

Note that branches are still open for bug fix backports, but NOT
supported or maintained by the project teams themselves (see details
in TC resolution [1]).

[1] https://governance.openstack.org/tc/resolutions/20230724-unmaintained-branches.html
[2] https://review.opendev.org/q/topic:vwx-unmaintained

Change-Id: I2444a5b4fcc3af4039e604a1e835a2ed5ada16fc