Release adjutant-ui RC1 for 2024.1 Caracal
This proposed a RC1 release for adjutant-ui to get any merged changes
out prior to the RC1 deadline for 2024.1 Caracal going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by afte Thursday, March 14th.

Change-Id: Ib916ac87b3b1bca5471a5590aa132900c6f90c22
Signed-off-by: Elod Illes <elod.illes@est.tech>