Release designate-dashboard RC1 for 2024.1 Caracal
This proposed a RC1 release for designate-dashboard to get any merged changes
out prior to the RC1 deadline for 2024.1 Caracal going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by after Thursday, March 14th.

Change-Id: I14739aef520d8c3e50a4cf023116f93d3b41feaf
Signed-off-by: Elod Illes <elod.illes@est.tech>