Release venus-dashboard RC1 for 2024.1 Caracal
This proposed a RC1 release for venus-dashboard to get any merged changes
out prior to the RC1 deadline for 2024.1 Caracal going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by after Thursday, March 14th.

Change-Id: I297a32cf65e8c63d33a5347dbec70d2c1084cb0c
Signed-off-by: Elod Illes <elod.illes@est.tech>