Release vitrage RC1 for 2024.1 Caracal
This proposed a RC1 release for vitrage to get any merged changes
out prior to the RC1 deadline for 2024.1 Caracal going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by after Thursday, March 14th.

Change-Id: Ief51d5df360ac508d129c3e73a38d1365fea1706
Signed-off-by: Elod Illes <elod.illes@est.tech>