Update for Unmaintained transition
We no longer have branches in extended maintenance state, these have
been switched to Unmaintained (eom). So we need to check for the latter
state when listing branches that could be deleted since the are EOLed.

Add oslo.utils to requirements since it is now needed by the
tools/delete_stable_branch.py script.

Change-Id: Ib948befa6f2706dc5dc50009b021af3a1bb389a8