Release glance-tempest-plugin for 2024.1 Caracal
This patch proposes a release for glance-tempest-plugin
to get the latest merged changes out prior to the final
2024.1 Caracal going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, March 21.

Change-Id: I9489911c5f6dc84704c3de83ccfaae00ed5015c6
Signed-off-by: Elod Illes <elod.illes@est.tech>