Release watcher-tempest-plugin for 2024.1 Caracal
This patch proposes a release for watcher-tempest-plugin
to get the latest merged changes out prior to the final
2024.1 Caracal going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, March 21.

Change-Id: I1bc78917e29885feb1d0b0f119fd994454663b8b
Signed-off-by: Elod Illes <elod.illes@est.tech>