Final RC patch for ovn-bgp-agent
This creates a new RC before the final release candidate deadline to
release the latest fixes and/or translations.

Please ack with a +1, or if there is some reason to hold off, leave a -1.

Patches with no response will be merged by Friday, March 29th.

$ git log --oneline --no-merges 2.0.0.0rc1..047d261
047d261 Fix placement of lsp when external_ids not in sync
9f93e32 Update TOX_CONSTRAINTS_FILE for stable/2024.1
abf1696 Update .gitreview for stable/2024.1

Change-Id: I673c43345527a8647586fba39a92b55bb6c8850c
Signed-off-by: Hervé Beraud <hberaud@redhat.com>