release whitebox-tempest-plugin
The whitebox-tempest-plugin is a deliverable of the qa team
but it is primarly consumed directly via git.

there has been a request to do a packaged release of it for rdo
and this patch just creates it.

https: //review.rdoproject.org/r/c/openstack/whitebox-tempest-plugin-distgit/+/52480

this change also addes the missing git shas for 0.0.1 and 0.0.2

Change-Id: I562103e4972006a47cf9232c8de2dbd54f5aa518