Add process step to check openstack map
During Caracal release it was discovered late that the openstack map
could use a refresh ahead of release communications. Since we should (in
theory) know the contents of the release (in terms of deliverables)
after milestone-2, add a step in the process to have a look at the map
and submit any necessary changes well ahead of the release week craze.

Change-Id: I51674b19afce0ea7b32f93ee989dbbf9df926f90