[storlets] Transition Zed to Unmaintained
This branch reached the end of its official stable maintained period,
hence this patch transition stable/zed branch to Unmaintained. (For
details about Unmaintained state please see: [1][2])

The transitions is as follows:
- tip of the stable/zed branch is tagged with zed-eom tag
- unmaintained/zed branch will be cut from zed-eom tag
- stable/zed branch will be deleted

Note:
- there will no longer be official releases off of zed branch after this
  transition
- backports can be pushed to unmaintained/yoga branch
- openstack-unmaintained-core (or <project>-unmaintained-core groups if
  such exists) have the rights to merge patches

Important: please +1 if the team is ready for us to proceed with this
transition, or -1 if there are any final Zed stable release with
backports currently in flight that we should wait for. For the latter
case, please propose a final Zed stable release patch, and rebase this
patch on top of that, with the latest commit hash of stable/zed.

The transition deadline is May 2nd, 2024. [3]

[1] https://docs.openstack.org/project-team-guide/stable-branches.html#unmaintained
[2] https://governance.openstack.org/tc/resolutions/20230724-unmaintained-branches.html
[3] https://releases.openstack.org

Change-Id: I41690e5591f55de5880744cebe92881728322f45