Set Zed state to Unmaintained
Zed series reached it's end of maintained state with the planned date
of May 2nd, 2024.

As most of the projects are ready with the transition [1], this patch
sets Zed as unmaintained.

Note that stable/zed branches are replaced with unmaintained/zed
branches and they are still open for bug fix backports, but NOT
supported or maintained by the project teams themselves (for details
about the Unmaintained state, see: [2][3]).

[1] https://review.opendev.org/q/topic:zed-unmaintained
[2] https://governance.openstack.org/tc/resolutions/20230724-unmaintained-branches.html
[3] https://docs.openstack.org/project-team-guide/stable-branches.html

Change-Id: I4b84d1ba03703f51897173b2e573c2e497f4b048