[blazar-tempest-plugin] Tag zed-last
Zed branch is moving to unmaintained state now[1]. Tempest and tempest
plugins are branchless which means master version of Tempest and
its plugins is used to test the supported stable branches.

Once any stable branch is moved to unmaintained state then, Tempest and
its plugins compatible tag[2][3] needs to be released which can be used
to test the unmaintained stable branches.

Tagging 'zed-last' as well as a new version with latest hash.

[1] https://review.opendev.org/q/topic:%22zed-unmaintained%22
[2] https://docs.openstack.org/tempest/latest/stable_branch_support_policy.html
[3] https://docs.openstack.org/tempest/latest/tempest_and_plugins_compatible_version_policy.html

Change-Id: I03c68c5b8bdd00abbc57aa1e1347834674c6899f