Release os-vif for Dalmatian-1 milestone
This is a library release for os-vif for the Dalmatian-1
milestone. This repository contains commits that have not been releases
for this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

Change-Id: Ibbc1bb15df23fac392aceb90a7f48defa80ba289
Signed-off-by: Elod Illes <elod.illes@est.tech>