Release python-cloudkittyclient for Dalmatian-1 milestone
This is a library release for python-cloudkittyclient for the Dalmatian-1
milestone. This repository contains commits that have not been releases
for this cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

Change-Id: I293f71412a3d3c04bd3a6ddb70fef113d48d200c
Signed-off-by: Elod Illes <elod.illes@est.tech>