Update team exceptions in ACL issues script
Update the team exceptions section in the aclissues.py detection script
so that it reflects that ironic and Puppet OpenStack teams have been
exempted from strict release management rules, together with references
to the changes that introduced the exemptions in the first place.

With this change, the aclissues.py script no longer reports false
negatives in ACLs for deliverables from those teams.

Change-Id: I63dddeb04dec129cbc1fffd9c638a1f35708aafb