Release oslo.config for Dalmatian-2 milestone
This is the Dalmatian-2 milestone release for oslo.config.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If the team is happy with the current release hash, please +1 this
patch and we will process the release. If you need to get some
important changes merged first, leave a -1 and update the patch with
a new commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed after July 4th.

Change-Id: I4f3b9af72b262c563c4e406a1c9fa184d317d725
Signed-off-by: Elod Illes <elod.illes@est.tech>