Release barbican for stable/2023.1
As a preparation for the transition of stable/2023.1 to Unmaintained,
this release picks up new commits to barbican since
the last release from stable/2023.1.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 16.0.0..9c8fb8c3
9c8fb8c3 Update secret:delete policy to allow admin to delete secret
2f3df026 Remove System scope from policy
b38b2139 Fix functional tests
e2bd3b95 Stable-only: Remove TripleO job
b5c2f838 Update TOX_CONSTRAINTS_FILE for stable/2023.1
441f7d47 Update .gitreview for stable/2023.1

Change-Id: Ie9aa4142097f227aa6b2b7feade7a5cf82ffc005
Signed-off-by: Elod Illes <elod.illes@est.tech>