Release blazar for stable/2023.1
As a preparation for the transition of stable/2023.1 to Unmaintained,
this release picks up new commits to blazar since
the last release from stable/2023.1.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 11.0.0..aa2451e
aa2451e Add support for specifying a custom CA bundle
24ca970 Update TOX_CONSTRAINTS_FILE for stable/2023.1
561a755 Update .gitreview for stable/2023.1

Change-Id: I16f48a72e1d741bcfead25facea8b50be8abace6
Signed-off-by: Elod Illes <elod.illes@est.tech>