Release cloudkitty for stable/2023.1
As a preparation for the transition of stable/2023.1 to Unmaintained,
this release picks up new commits to cloudkitty since
the last release from stable/2023.1.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 18.0.0..00f7ce2
dd5bb9e Fix devstack runprocess for cloudkitty api
af1019d Fix API report requests when using opensearch
1cac176 Add OpenSearch as a v2 storage backend
4592836 Update Monasca deprecation and removal details
8c44864 Fix retrieval of reprocessing tasks
08a22b5 Fix random unit test failures
dd5c9fd Fix formatting in release note
073641f Update TOX_CONSTRAINTS_FILE for stable/2023.1
aabbc2e Update .gitreview for stable/2023.1

Change-Id: I838e1e9d8079cb2b1752de76bdd99cee14690051
Signed-off-by: Elod Illes <elod.illes@est.tech>