Release ironic for stable/2023.1
As a preparation for the transition of stable/2023.1 to Unmaintained,
this release picks up new commits to ironic since
the last release from stable/2023.1.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 21.4.1..072e02da3
072e02da3 fix: Fix class typo for portgroup. Portgroup instead of PortGroup
86358c89e ci: stable-only: explicitly pin centos build
1060a570b Inject a randomized publisher id

Change-Id: Icc9f286fd00cb8fe738a16247fca4c9ab418e2c0
Signed-off-by: Elod Illes <elod.illes@est.tech>