Release keystone for stable/2023.1
As a preparation for the transition of stable/2023.1 to Unmaintained,
this release picks up new commits to keystone since
the last release from stable/2023.1.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 23.0.1..c725173cf
c725173cf Remove reference to devstack-gate
c81bc1b29 Add domain scoping to list_domains
b6c20d912 Allow domain admin to view roles
17da229f0 Allow domain users to manage credentials
b449d86ee Allow admin to access tokens and credentials
9b2f7aa4b Add ability to create users and projects from keystone-manage
9ec8e7cca Fix old arm64 job template
ee9a96672 Remove unused old job templates and experimental jobs
828519f46 Normalize policy checks for domain-scoped tokens
1c2d51788 Allow users with "admin" role to get projects
f7892ce2c Fix policies for groups
658fd7d58 Consistent and Secure RBAC (Phase 1)
ea8c8aa98 Don't forget to check if authorization fails
a9d5e7eea Add an option to randomize LDAP urls list
c1247a6a3 docs: Clarify lack of LDAP assignment back end

Change-Id: I5d8adff6569af0361fceaf31b069cf72c9b69c17
Signed-off-by: Elod Illes <elod.illes@est.tech>