Release keystoneauth for stable/2023.1
As a preparation for the transition of stable/2023.1 to Unmaintained,
this release picks up new commits to keystoneauth since
the last release from stable/2023.1.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 5.1.2..e96c503
e96c503 Make v3oidcpassword send client_id
97c2d0b Update TOX_CONSTRAINTS_FILE for stable/2023.1
8aea357 Update .gitreview for stable/2023.1

Change-Id: Iedd7860c5906e41c148848fe40043eb92ead3eb1
Signed-off-by: Elod Illes <elod.illes@est.tech>