Release magnum for stable/2023.1
As a preparation for the transition of stable/2023.1 to Unmaintained,
this release picks up new commits to magnum since
the last release from stable/2023.1.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 16.0.2..77293ee6
77293ee6 Fix oslo policy file genrator tool for Magnum
6d05cd31 Update chart.metadata.version to reflect breaking change in helm v3.5.2

Change-Id: I00b947bccac86dc8605df6dac5401b5ef4cb552a
Signed-off-by: Elod Illes <elod.illes@est.tech>