Release manila-ui for stable/2023.1
As a preparation for the transition of stable/2023.1 to Unmaintained,
this release picks up new commits to manila-ui since
the last release from stable/2023.1.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 9.0.0..ea689f5
ea689f5 Avoid server error caused by UnicodeDecodeError
90e74cc Imported Translations from Zanata
b9415b7 Update TOX_CONSTRAINTS_FILE for stable/2023.1
f4ddeb7 Update .gitreview for stable/2023.1

Change-Id: Ie92aba5b78c95ba152d8717e02251a402c1e9787
Signed-off-by: Elod Illes <elod.illes@est.tech>