Release manila for stable/2023.1
As a preparation for the transition of stable/2023.1 to Unmaintained,
this release picks up new commits to manila since
the last release from stable/2023.1.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 16.2.0..dad310243
590901073 [ci] Fix zuul configuration errors
6dad1afca bug fix: error message fix
f4b8a7bea Conditional Import for FIPS Compliance
0e3523e11 Fixed bug 2069125: Manila driver error with ONTAP SVM-scoped user
18b6562dd Fix leak of ports on share server deletion
fc5c241a6 Forbid resetting state of active replicas
7dbba77f8 Retry on connection error to neutron

Change-Id: Iaa6b61adff149fdd4e42ed19fb188c82c6bfaaf5
Signed-off-by: Elod Illes <elod.illes@est.tech>