Release metalsmith for stable/2023.1
As a preparation for the transition of stable/2023.1 to Unmaintained,
this release picks up new commits to metalsmith since
the last release from stable/2023.1.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 1.10.0..b3787f9
b3787f9 Set role default metalsmith_provisioning_timeout
c71ca5f Update TOX_CONSTRAINTS_FILE for stable/2023.1
7bf0996 Update .gitreview for stable/2023.1

Change-Id: Ic75ea173723b5a1aa52e8edc927992457b4291ae
Signed-off-by: Elod Illes <elod.illes@est.tech>