Release neutron for stable/2023.1
As a preparation for the transition of stable/2023.1 to Unmaintained,
this release picks up new commits to neutron since
the last release from stable/2023.1.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 22.2.0..a1732d4f39
a1732d4f39 Don't print traceback if standard attr is missing on update

Change-Id: I82dafb3520f9c286e557705601caf5e2eb38778c
Signed-off-by: Elod Illes <elod.illes@est.tech>