Release octavia-dashboard for stable/2023.1
As a preparation for the transition of stable/2023.1 to Unmaintained,
this release picks up new commits to octavia-dashboard since
the last release from stable/2023.1.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 11.0.0..7eaa562
6f990e0 Fix popover attributes in member add dialog
fa8fbcf Fix error related to barbicanclient
d1a6dd8 Update TOX_CONSTRAINTS_FILE for stable/2023.1
ce03362 Update .gitreview for stable/2023.1

Change-Id: I86ca65689383d431212c729b9772979e3f32a7e3
Signed-off-by: Elod Illes <elod.illes@est.tech>