Release openstacksdk for stable/2023.1
As a preparation for the transition of stable/2023.1 to Unmaintained,
this release picks up new commits to openstacksdk since
the last release from stable/2023.1.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 1.0.1..eee309c58
eee309c58 Remove retired project Senlin job
36af2dfb1 Fix list of server migrations
f57165713 Do not pin openstacksdk to master
078b34c79 Update TOX_CONSTRAINTS_FILE for stable/2023.1
e96ad1303 Update .gitreview for stable/2023.1

Change-Id: I50343d4abc90096506c4af2266aa4829bd0e8773
Signed-off-by: Elod Illes <elod.illes@est.tech>