Release os-brick for stable/2023.1
As a preparation for the transition of stable/2023.1 to Unmaintained,
this release picks up new commits to os-brick since
the last release from stable/2023.1.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 6.2.3..78609b6
78609b6 Silence warning when running in a container with overlayfs
c526c3b Fix: FC partial target scan
3c570e8 Dell Powerflex: Add new VOLUME_NOT_MAPPED_ERROR

Change-Id: Ic29d4199e535b60fd6f3d2c84ea38ca95fbac7c2
Signed-off-by: Elod Illes <elod.illes@est.tech>