Release os-ken for stable/2023.1
As a preparation for the transition of stable/2023.1 to Unmaintained,
this release picks up new commits to os-ken since
the last release from stable/2023.1.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 2.6.0..6b427e4b
6b427e4b Raise ValueError in case unpack_from returns zero length
4df22c4a Update TOX_CONSTRAINTS_FILE for stable/2023.1
a0cb802c Update .gitreview for stable/2023.1
3ae818b9 Docs build: Add requirements.txt to deps

Change-Id: If81d1590c13cc647ef0547f1909578a1e76ec794
Signed-off-by: Elod Illes <elod.illes@est.tech>