Release oslo.db for stable/2023.1
As a preparation for the transition of stable/2023.1 to Unmaintained,
this release picks up new commits to oslo.db since
the last release from stable/2023.1.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 12.3.2..be0515d
be0515d exc_filters: Handle OperationalError for MariaDB/Galera

Change-Id: I075f3ef47ba5c3dec1341bb5cdf49d24ef22700f
Signed-off-by: Elod Illes <elod.illes@est.tech>