Release python-glanceclient for stable/2023.1
As a preparation for the transition of stable/2023.1 to Unmaintained,
this release picks up new commits to python-glanceclient since
the last release from stable/2023.1.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 4.3.1..3a2ef8b
3a2ef8b Bump the CHUNKSIZE to use CPU more efficiently

Change-Id: I26572fad20dea07c5a38bf54f121fe42c7f00a91
Signed-off-by: Elod Illes <elod.illes@est.tech>