Release python-manilaclient for stable/2023.1
As a preparation for the transition of stable/2023.1 to Unmaintained,
this release picks up new commits to python-manilaclient since
the last release from stable/2023.1.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 4.4.1..96e51a3
96e51a3 [OSC] Fix output format for osc share subnet create
9708c1b Don't set availability_zone for 2.32 version of api
39021ab Fix share force delete case
d229910 Fix "per_share_gigabytes" in "openstack quota set"

Change-Id: I17585b8b914f85d410f8cd85d901cb03a892548e
Signed-off-by: Elod Illes <elod.illes@est.tech>