Release python-openstackclient for stable/2023.1
As a preparation for the transition of stable/2023.1 to Unmaintained,
this release picks up new commits to python-openstackclient since
the last release from stable/2023.1.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 6.2.0..02172959
02172959 [stable-only] CI: Drop all tips jobs
a03e3dbf Fix "access rule" commands to only use ID
07c5a26f Fix pep issue in the network service provider
06315b93 Update TOX_CONSTRAINTS_FILE for stable/2023.1
4e9f6387 Update .gitreview for stable/2023.1

Change-Id: Ie032979a3c67472f16d9f2acacbe2c40e44f3f78
Signed-off-by: Elod Illes <elod.illes@est.tech>