Release trove for stable/2023.1
As a preparation for the transition of stable/2023.1 to Unmaintained,
this release picks up new commits to trove since
the last release from stable/2023.1.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

$ git log --oneline --no-merges 19.0.0..6be287cb
3f08f374 remove .coverage file from repository
68f6bfa0 Fix error can't detach the configuration group binding
2b6ea5e2 Retire TripleO: remove the TripleO jobs/ref
363c1df7 stable-only: Pin bandit
9dd308f5 Fix guest agent failed to use swiftclient over TLS
bb8bbbeb Add region name to trove-guestagent
53eef3bc Change settings for stable/2023.1
1e7bfafd [CI]: fix trove-tempest jobs on stable/2023.1
959e9cd5 [CI]: Don't run periodic jobs on stable branchs
f62243a9 Fix "create database" failed for postgress
fee50052 [CI] Don't run public trove images jobs in periodic pipeline
edbdbe36 Fixes permission problem when restoring backup
8d54913b Update TOX_CONSTRAINTS_FILE for stable/2023.1
e6728c1d Update .gitreview for stable/2023.1

Change-Id: I9badf2b4fc7b4f161a0b0cb1abb619fb66c8b927
Signed-off-by: Elod Illes <elod.illes@est.tech>