Release final ceilometermiddleware for 2024.2 (Dalmatian)
This proposes a release for ceilometermiddleware to get any final merged changes
out prior to the lib freeze for 2024.2 going into effect.
If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

$ git log --oneline --no-merges 3.4.0..dec1817
dec1817 If-else with same path in both directions
76ee310 Remove old excludes
d937ebe reno: Update master for unmaintained/zed

Change-Id: I2568e796f1890c32329f164e2d4ad3c786fe5b48
Signed-off-by: Dr. Jens Harbott <harbott@osism.tech>