Release final heat-translator for 2024.2 (Dalmatian)
This proposes a release for heat-translator to get any final merged changes
out prior to the lib freeze for 2024.2 going into effect.
If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

$ git log --oneline --no-merges 3.0.0..fa90abe
2c32960 Use python 3.10 in unit tests with latest tosca-parser
8a96764 Drop explicit installation of extras

Change-Id: I4e16588eed86ff6b18f6fbc71ec407004dfed205
Signed-off-by: Dr. Jens Harbott <harbott@osism.tech>