Release final ironic-lib for 2024.2 (Dalmatian)
This proposes a release for ironic-lib to get any final merged changes
out prior to the lib freeze for 2024.2 going into effect.
If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

$ git log --oneline --no-merges 6.1.0..1ca3c8c
1ca3c8c Fix invalid UTF-8 characters in execute output
c3d59df Update to match latest development cycle
24b766f Fix codespell reported errors

Change-Id: I600b99fa6bf840ffcf43a488cf5425de6c23d215
Signed-off-by: Dr. Jens Harbott <harbott@osism.tech>