Release final keystoneauth for 2024.2 (Dalmatian)
This proposes a release for keystoneauth to get any final merged changes
out prior to the lib freeze for 2024.2 going into effect.
If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

$ git log --oneline --no-merges 5.7.0..cca6c92
0b82d42 Update requests of device flow for public client
d552a9a Add OTP to v3OIDCpassword plugin
a0c9d5e Remove old excludes

Change-Id: I7be922327eba1e53d79ded3f8ed28f6e6ad0f0dd
Signed-off-by: Dr. Jens Harbott <harbott@osism.tech>