Release final keystonemiddleware for 2024.2 (Dalmatian)
This proposes a release for keystonemiddleware to get any final merged changes
out prior to the lib freeze for 2024.2 going into effect.
If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

$ git log --oneline --no-merges 10.7.0..4c001ba
c890d93 reno: Update master for unmaintained/zed
8ca2956 Remove old excludes

Change-Id: I43e059ee2f7909fbf9f34d43f3bdb6f427d156e8
Signed-off-by: Dr. Jens Harbott <harbott@osism.tech>