Release final kuryr for 2024.2 (Dalmatian)
This proposes a release for kuryr to get any final merged changes
out prior to the lib freeze for 2024.2 going into effect.
If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

$ git log --oneline --no-merges 3.0.0..84c22b9
84c22b9 reno: Update master for unmaintained/zed
60e7f0c reno: Update master for unmaintained/xena
41f4892 reno: Update master for unmaintained/wallaby
44bc968 reno: Update master for unmaintained/victoria
41c8d40 Update master for stable/2024.1
ad96ea6 reno: Update master for unmaintained/yoga

Change-Id: I3aae52dbdc6724b9d13105580f31538104474251
Signed-off-by: Dr. Jens Harbott <harbott@osism.tech>