Release final mistral-extra for 2024.2 (Dalmatian)
This proposes a release for mistral-extra to get any final merged changes
out prior to the lib freeze for 2024.2 going into effect.
If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

$ git log --oneline --no-merges 13.2.0..660dd14
7b9e415 reno: Update master for unmaintained/xena
a504b98 reno: Update master for unmaintained/wallaby
86ebf67 reno: Update master for unmaintained/victoria
b225790 Update master for stable/2024.1
463e867 reno: Update master for unmaintained/yoga

Change-Id: I85564877f4e0d038539eb6229edb34554ae2b6d2
Signed-off-by: Dr. Jens Harbott <harbott@osism.tech>