Release final neutron-lib for 2024.2 (Dalmatian)
This proposes a release for neutron-lib to get any final merged changes
out prior to the lib freeze for 2024.2 going into effect.
If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

$ git log --oneline --no-merges 3.14.1..8fc2cc6f
d69eae32 Fix python modules docs build
551bcff7 Add API extension ``tag-creation``

Change-Id: I0b6e3f6e261f2da05b5e7db715e08fe04abea91b
Signed-off-by: Dr. Jens Harbott <harbott@osism.tech>