Release final os-ken for 2024.2 (Dalmatian)
This proposes a release for os-ken to get any final merged changes
out prior to the lib freeze for 2024.2 going into effect.
If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

$ git log --oneline --no-merges 2.9.0..3939487d
0f453910 Remove old excludes
c48dbd60 Add periodic weekly job for os-ken
8d63e980 reno: Update master for unmaintained/zed
ea4fe8fa Update python classifier in setup.cfg

Change-Id: I45f5011968f9f474d2d4dc7a50100cc17a358ed1
Signed-off-by: Dr. Jens Harbott <harbott@osism.tech>