Release final os-vif for 2024.2 (Dalmatian)
This proposes a release for os-vif to get any final merged changes
out prior to the lib freeze for 2024.2 going into effect.
If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

$ git log --oneline --no-merges 3.6.0..bb3e3da
b11239f Remove old excludes

Change-Id: Ic034923d8d74ca2358e093915adfccda7ef2e922
Signed-off-by: Dr. Jens Harbott <harbott@osism.tech>