Release final stevedore for 2024.2 (Dalmatian)
This proposes a release for stevedore to get any final merged changes
out prior to the lib freeze for 2024.2 going into effect.
If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

$ git log --oneline --no-merges 5.2.0..51134a4
51134a4 reno: Update master for unmaintained/zed
8ab9244 Remove old excludes
82186ba Update master for stable/2024.1
41342e6 reno: Update master for unmaintained/xena
199ea51 reno: Update master for unmaintained/wallaby
f2997c4 reno: Update master for unmaintained/victoria

Change-Id: I0cc7935705c8fab9d8ec8e42258bc6874c2b5aec
Signed-off-by: Dr. Jens Harbott <harbott@osism.tech>