Release final osc-placement for 2024.2 Dalmatian
This is a release for osc-placement to get any merged but
unreleased changes out prior to client library freeze for
2024.2 Dalmatian going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, August 29th.

$ git log --oneline --no-merges 4.4.0..2a3922d
2a3922d Remove use of distutils

Change-Id: Ia2ffee18aedd1c73290e03131bee47d5d4717aa7
Signed-off-by: Elod Illes <elod.illes@est.tech>