Release final python-aodhclient for 2024.2 Dalmatian
This is a release for python-aodhclient to get any merged but
unreleased changes out prior to client library freeze for
2024.2 Dalmatian going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, August 29th.

$ git log --oneline --no-merges 3.5.1..5738d4c
70ed8b2 Remove old excludes
169938f reno: Update master for unmaintained/zed
8a3f259 reno: Update master for unmaintained/wallaby
03eab3c Update master for stable/2024.1

Change-Id: I0b6e97f10d0b5a45081dc759dab141fd2dcc3968
Signed-off-by: Elod Illes <elod.illes@est.tech>