Release final python-blazarclient for 2024.2 Dalmatian
This is a release for python-blazarclient to get any merged but
unreleased changes out prior to client library freeze for
2024.2 Dalmatian going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, August 29th.

$ git log --oneline --no-merges 4.0.1..e1f6650
e1f6650 Bump hacking to 7.0.0 and fix linting issues
1332265 Support flavor-based instance reservation
7a3a6a8 reno: Update master for unmaintained/zed
4a7c0e9 Update master for stable/2024.1
ee25d60 reno: Update master for unmaintained/xena
d9cc3a5 reno: Update master for unmaintained/wallaby
9e2684e reno: Update master for unmaintained/victoria

Change-Id: I59f991dce8342f35c0ecf4fed7d8623bd7849770
Signed-off-by: Elod Illes <elod.illes@est.tech>