Release final python-brick-cinderclient-ext for 2024.2 Dalmatian
This is a release for python-brick-cinderclient-ext to get any merged but
unreleased changes out prior to client library freeze for
2024.2 Dalmatian going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, August 29th.

$ git log --oneline --no-merges 2.3.0..fbefcb4
fbefcb4 reno: Update master for unmaintained/zed
27474b1 Update CI for Dalmatian
27f39e3 Update master for stable/2024.1
02fdfb5 reno: Update master for xena Unmaintained status
e537a4d reno: Update master for wallaby Unmaintained status
e910e17 reno: Update master for victoria Unmaintained status

Change-Id: I8b0d44aa3fdba3a6382674677a1ae3dfa98f53d0
Signed-off-by: Elod Illes <elod.illes@est.tech>