Release final python-designateclient for 2024.2 Dalmatian
This is a release for python-designateclient to get any merged but
unreleased changes out prior to client library freeze for
2024.2 Dalmatian going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, August 29th.

$ git log --oneline --no-merges 6.0.1..4f98ae3
4f98ae3 reno: Update master for unmaintained/zed
ea4128a Update master for stable/2024.1
ec46cc6 reno: Update master for unmaintained/xena
e5d2a21 reno: Update master for unmaintained/wallaby
9b40bd7 reno: Update master for unmaintained/victoria
7787fef Bump hacking

Change-Id: I5fed60b707417fe79353a103f6b7a2503a365de9
Signed-off-by: Elod Illes <elod.illes@est.tech>