Release final python-ironic-inspector-client for 2024.2 Dalmatian
This is a release for python-ironic-inspector-client to get any merged but
unreleased changes out prior to client library freeze for
2024.2 Dalmatian going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, August 29th.

$ git log --oneline --no-merges 5.1.0..5e23849
5e23849 Deprecate the project
cd9416b reno: Update master for unmaintained/zed
36204a2 Remove old excludes
37b6fdc reno: Update master for unmaintained/xena
6bc403e reno: Update master for unmaintained/wallaby
daa819a reno: Update master for unmaintained/victoria
8a908f7 Update master for stable/2024.1

Change-Id: I71c7310bdc11b4715cde2ebfa73ddedc09bf17c7
Signed-off-by: Elod Illes <elod.illes@est.tech>