Release final python-keystoneclient for 2024.2 Dalmatian
This is a release for python-keystoneclient to get any merged but
unreleased changes out prior to client library freeze for
2024.2 Dalmatian going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, August 29th.

$ git log --oneline --no-merges 5.4.0..a6356495
2fe440b7 reno: Update master for unmaintained/zed
60f89caa Remove old excludes
cf981fac reno: Update master for unmaintained/xena
67c56423 reno: Update master for unmaintained/wallaby
a24415e8 reno: Update master for unmaintained/victoria
07b7ddde Update master for stable/2024.1
916814ec tox: Drop envdir

Change-Id: Id90e04170ef3adb7bf9545a8da302d21c746f45d
Signed-off-by: Elod Illes <elod.illes@est.tech>