Release final python-magnumclient for 2024.2 Dalmatian
This is a release for python-magnumclient to get any merged but
unreleased changes out prior to client library freeze for
2024.2 Dalmatian going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, August 29th.

$ git log --oneline --no-merges 4.6.0..6fa75fe
3833c46 Refactor TLS cert fetching for config

Change-Id: I87f6936edd9667a0aae02e1ad4acede9cd6944eb
Signed-off-by: Elod Illes <elod.illes@est.tech>