Release final python-mistralclient for 2024.2 Dalmatian
This is a release for python-mistralclient to get any merged but
unreleased changes out prior to client library freeze for
2024.2 Dalmatian going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, August 29th.

$ git log --oneline --no-merges 5.2.0..926301e
926301e Update master for stable/2024.1

Change-Id: I305681f5b2647618aae420de9cee6cdb1caccb8a
Signed-off-by: Elod Illes <elod.illes@est.tech>