Release final python-novaclient for 2024.2 Dalmatian
This is a release for python-novaclient to get any merged but
unreleased changes out prior to client library freeze for
2024.2 Dalmatian going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, August 29th.

$ git log --oneline --no-merges 18.6.0..2bd135c1
2bd135c1 Remove old excludes
ec292ac6 reno: Update master for unmaintained/zed
3ffc5e5b Update master for stable/2024.1
99bbecfc reno: Update master for unmaintained/xena
f2d51a0f reno: Update master for unmaintained/wallaby
59a147ab reno: Update master for unmaintained/victoria

Change-Id: Ie4a1fe2fd328a1399bc31b327eed2d074aa6271e
Signed-off-by: Elod Illes <elod.illes@est.tech>