Release final python-octaviaclient for 2024.2 Dalmatian
This is a release for python-octaviaclient to get any merged but
unreleased changes out prior to client library freeze for
2024.2 Dalmatian going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, August 29th.

$ git log --oneline --no-merges 3.7.0..59154ff
ab78ca7 Add SR-IOV support for members
2820fbd reno: Update master for unmaintained/zed
b9d8f34 Update master for stable/2024.1
4802100 Support filtering amphora list by image ID
282ee6a reno: Update master for unmaintained/xena
4b2f3ba reno: Update master for unmaintained/wallaby
06db319 reno: Update master for unmaintained/victoria

Change-Id: I18e81815364cde733a1e7ebad9952b841429e9b2
Signed-off-by: Elod Illes <elod.illes@est.tech>