Release final python-vitrageclient for 2024.2 Dalmatian
This is a release for python-vitrageclient to get any merged but
unreleased changes out prior to client library freeze for
2024.2 Dalmatian going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, August 29th.

Change-Id: Id94792ba87e1a64f1d5b38d9966b454eb0d9bf78
Signed-off-by: Elod Illes <elod.illes@est.tech>