Release final python-zaqarclient for 2024.2 Dalmatian
This is a release for python-zaqarclient to get any merged but
unreleased changes out prior to client library freeze for
2024.2 Dalmatian going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, August 29th.

Change-Id: Ia9f9e7f0a5f9e5ea3d6f5d6d7f970faff8ddd1dc
Signed-off-by: Elod Illes <elod.illes@est.tech>