Release final python-zunclient for 2024.2 Dalmatian
This is a release for python-zunclient to get any merged but
unreleased changes out prior to client library freeze for
2024.2 Dalmatian going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, August 29th.

$ git log --oneline --no-merges 5.0.0..0af6771
288bb2e reno: Update master for unmaintained/xena
6641386 reno: Update master for unmaintained/wallaby
fd134ee reno: Update master for unmaintained/victoria
5286ee7 Update master for stable/2024.1
cd540e7 reno: Update master for unmaintained/yoga

Change-Id: I76ac0b98c110200bf7b6df14b86e98ea22556168
Signed-off-by: Elod Illes <elod.illes@est.tech>