Release ansible-role-lunasa-hsm RC1 for 2024.2 Dalmatian
This proposed a RC1 release for ansible-role-lunasa-hsm to get any merged changes
out prior to the RC1 deadline for 2024.2 Dalmatian going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, September 13th.

Change-Id: I58773dc8ca1a5cfc18ef76d97d5b7b1a480b2279
Signed-off-by: Elod Illes <elod.illes@est.tech>