Release nova RC1 for 2024.2 Dalmatian
This proposed a RC1 release for nova to get any merged changes
out prior to the RC1 deadline for 2024.2 Dalmatian going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged by Friday, September 13th.

Change-Id: I1ae117d460fca058e367c68ee5be047d8c4a2dfa
Signed-off-by: Elod Illes <elod.illes@est.tech>