Remove description to simplify release-announce
Current release-announce job proceeds with installing the package being
released, just in order to compute the "description" of the package and
include it in the announcement. This is costly and error-prone, for very
little gain.

This change removes the costly computation of the "description". The
"project_name" was overridden by $ZUUL_PROJECT anyway. The email
template already has logic to skip the description if it's not provided.

Change-Id: I5a60556f7d321fd4c37f1424509469da28a1eec9