stable/2024.2 branches for Ironic deliverables
Create stable/2024.2 branches for ironic team deliverables that were
still missing one, based on latest available dalmatian releases.

If the ironic team wants to wait and make another release for any of
those before the branches are cut, please -1 this patch to have it held,
or update that patch to include another release and stable branch point.

Please +1 the patch if this is ready to go. If no feedback is given at
all, this patch will be merged on Friday, September 20.

Change-Id: If2a37acac48a9ce613e48d350000c312f8bfbb36