Release heat-tempest-plugin for 2024.2 Dalmatian
This patch proposes a release for heat-tempest-plugin
to get the latest merged changes out prior to the final
2024.2 Dalmatian going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, September 19th.

Change-Id: I7c506a8ea0a1be151128c18b69d4d37c8ee80a9d
Signed-off-by: Elod Illes <elod.illes@est.tech>