Release octavia-tempest-plugin for 2024.2 Dalmatian
This patch proposes a release for octavia-tempest-plugin
to get the latest merged changes out prior to the final
2024.2 Dalmatian going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, September 19th.

Change-Id: If6568e0e48ed44a5ddc339cc1d7072c64195ad9b
Signed-off-by: Elod Illes <elod.illes@est.tech>