Release telemetry-tempest-plugin for 2024.2 Dalmatian
This patch proposes a release for telemetry-tempest-plugin
to get the latest merged changes out prior to the final
2024.2 Dalmatian going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, September 19th.

Change-Id: I9e7b0184fbe823251f7229a250b45201b861c921
Signed-off-by: Elod Illes <elod.illes@est.tech>