2024.2 Dalmatian final releases for cycle-with-rc projects
This tags the final release for all cycle-with-rc projects for 2024.2
Dalmatian. While not required, having PTLs ack this patch would be nice
to include that record in the merged metadata.

Release activity should now be frozen until this is processed on
October 2nd. Any bugfixes merged to stable/2024.2 can be queued up and
ready to go as a follow on stable release after the coordinated release
date.

There are different diff-start values for each project. This is used
when generating release notes to denote the point where the
stable/2024.1 branch started to diverge from what is included on
stable/2024.2. For some, this will be the final 2024.1 Caracal major
version, but if multiple RCs were done, this may show the initial RC1
release version. This is normal.

This patch will be merged on October 2nd.

Change-Id: I093fc9dda6d787ead5b40defccddadbf913046ec