[tool] Update listing tool for open and unreleased changes
The list_unreleased_and_open_changes.sh was not prepared to use
different branch name than the series name. Since 2023.1 Antelope
release we have a 'series name' (antelope) and a 'release ID' (2023.1)
so this patch updates the script to work properly and use the series
name when that is required and use release ID for branch name.

Change-Id: Ibd932d9b9e1c0f5d6e3269c6f1822d7b1d56c7be