Add --except-type to list-deliverables command
We have the --type argument, where we can list which type of
deliverables we want to be listed, but we don't have the opposite
approach where we list the types that we DON'T want to be listed.

This patch introduces the --except-type argument with wich we can
list the deliverables with excluding some types, for example:

$ list-deliverables --series epoxy --except-type tempest-plugin

The above command lists all deliverables from 2025.1 Epoxy series
that aren't tempest-plugins.

Change-Id: If00115d3e7de055da2bd7a2dcd08298f70bd5989